% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAbioticObservations.R
\name{addAbioticObservations}
\alias{addAbioticObservations}
\title{Add abiotic observation records}
\usage{
addAbioticObservations(target, x, projectTitle, mapping,
  measurementMethods = list(), missing.values = c(NA, ""), verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one plot observation. Columns can be varied.}

\item{projectTitle}{A string to identify the project title, which can be the same as one of the currently defined in \code{target}.}

\item{mapping}{A list with element names 'plotName', 'obsStartDate', used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Abiotic variables that can be mapped are: 'phosphorus', 'pottasium', 'magnesium', 'nitrogen' and 'pH'.
Additional optional mappings are: 'subPlotName' and 'obsEndDate'.}

\item{measurementMethods}{A named list of objects of class \code{\linkS4class{VegXMethod}} with the measurement method
for each of the abiotic variables stated in \code{mapping}. List names should be the same as abiotic variables
(e.g. \code{list(pH = pHmeth)} to specify the use of method '\code{pHmeth}' for pH measurements).}

\item{missing.values}{A character vector of values that should be considered as missing observations/measurements.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds abiotic observation records to a VegX object from a data table where rows are plot observations,
using a mapping to identify plot observation: plot, subplot (optional) and observation date.
Additional mappings are used to map specific abiotic variables.
}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
\code{\link{addTaxonObservations}}, \code{\link{addTreeObservations}}.
}
