% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCommunityObservations.R
\name{addCommunityObservations}
\alias{addCommunityObservations}
\title{Add community observation records}
\usage{
addCommunityObservations(
  target,
  x,
  mapping,
  methods = list(),
  date.format = "\%Y-\%m-\%d",
  missing.values = c(NA, ""),
  verbose = TRUE
)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one plot observation. Columns can be varied.}

\item{mapping}{A list with element names 'plotName', 'obsStartDate', used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Additional optional mappings are: 'subPlotName'.}

\item{methods}{A named list of objects of class \code{\linkS4class{VegXMethodDefinition}} with the measurement method
for each of the community measurements listed in \code{mapping}. List names should be the same as subject measurement variables.
Alternatively, methods can be specified using strings if predefined methods exist (see \code{\link{predefinedMeasurementMethod}}).}

\item{date.format}{A character string specifying the input format of dates (see \code{\link{as.Date}}).}

\item{missing.values}{A character vector of values that should be considered as missing observations/measurements.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds community observation records to a VegX object from a data table where rows are plot observations.
}
\details{
Missing value policy:
\itemize{
  \item{Missing 'plotName' or 'obsStartDate' values are interpreted as if the previous non-missing value has to be used to define plot observation.}
  \item{Missing 'subPlotName' values are interpreted in that observation refers to the parent plotName.}
  \item{Missing measurements are simply not added to the Veg-X document.}
}
}
\examples{
data(Mokihinui)

# Simulate measurement of basal area
moki_site$BA = pmax(0, rnorm(nrow(moki_site), 10, 5))

# Define mapping
mapping = list(plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", basal_area = "BA")


x = addCommunityObservations(newVegX(), moki_site, mapping = mapping,
                        methods = list(basal_area = "Basal area/m2*ha-1"))
                        
# Inspect the result
head(showElementTable(x, "communityObservation"))
}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other add functions: 
\code{\link{addAggregateOrganismObservations}()},
\code{\link{addIndividualOrganismObservations}()},
\code{\link{addPlotGeometries}()},
\code{\link{addPlotLocations}()},
\code{\link{addPlotObservations}()},
\code{\link{addSiteCharacteristics}()},
\code{\link{addSiteObservations}()},
\code{\link{addStratumObservations}()},
\code{\link{addSurfaceCoverObservations}()},
\code{\link{addTaxonBySiteData}()}
}
\concept{add functions}
