% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlotObservations.R
\name{addPlotObservations}
\alias{addPlotObservations}
\title{Add plot observation records}
\usage{
addPlotObservations(target, x, mapping, missing.values = c(NA, ""),
  verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one plot observation. Columns can be varied.}

\item{mapping}{A list with element names 'plotName', 'obsStartDate', used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Additional optional mappings are: 'projectTitle'  'obsEndDate', 'subPlotName'.}

\item{missing.values}{A character vector of values that should be considered as missing observations/measurements.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds plot observation records to a VegX object from a data table where rows are plot observations,
using a mapping to identify plot observation: plot, subplot (optional), observation start date and end date (optional).
}
\details{
Missing projectTitle, plotName, obsStartDate or obsEndDate values are interpreted as if the previous non-missing value has to be used to define plot observation.
Missing subPlotName values are interpreted in that observation refers to the parent plotName.
}
\examples{
# Load source data
data(mokihinui)

# Define mapping
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")

# Create a new Veg-X document with projects, plots and plot observations (no data)
x = addPlotObservations(newVegX(), site, mapping = mapping)

# Examine the result
showElementTable(x, "plot")
showElementTable(x, "plotObservation")

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other add functions: \code{\link{addAggregateOrganismObservations}},
  \code{\link{addIndividualOrganismObservations}},
  \code{\link{addPlotLocations}},
  \code{\link{addSiteCharacteristics}},
  \code{\link{addSiteObservations}},
  \code{\link{addStratumObservations}},
  \code{\link{addTaxonBySiteData}}
}
