% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSiteCharacteristics.R
\name{addSiteCharacteristics}
\alias{addSiteCharacteristics}
\title{Adds/replaces site information}
\usage{
addSiteCharacteristics(target, x, mapping, measurementMethods = list(),
  missing.values = c(NA, ""), verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one plot observation. Columns can be varied.}

\item{mapping}{A list with at least element name 'plotName', is used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Site variables that can be mapped are: 'slope', 'aspect', 'landform', 'parentMaterial'.
Additional optional mappings are: 'subPlotName'.}

\item{measurementMethods}{A named list of objects of class \code{\linkS4class{VegXMethodDefinition}} with the measurement method
for each of the abiotic variables stated in \code{mapping}. List names should be the same as abiotic variables
(e.g. \code{list(aspect = aspectMeth)} to specify the use of method '\code{aspectMeth}' for aspect measurements).}

\item{missing.values}{A character vector of values that should be considered as missing observations/measurements.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds/replaces static site characteristics (topography, geology, ...) to plot elements of a VegX object from a data table where rows are plots.
}
\details{
Missing value policy:
 \itemize{
    \item{Missing 'plotName' values are interpreted as if the previous non-missing value has to be used to define plot.}
    \item{Missing 'subPlotName' values are interpreted in that data refers to the parent plotName.}
    \item{Missing measurements (e.g. 'aspect', 'slope',...) are simply not added to the Veg-X document.}
 }
}
\examples{
# Load source data
data(mokihinui)

# Define mapping
sitemapping = list(plotName = "Plot", subPlotName = "Subplot",
                   slope = "PlotSlope", aspect = "PlotAspect")

# Define site methods
slopeDeg = predefinedMeasurementMethod("Slope/degrees")
aspectDeg = predefinedMeasurementMethod("Aspect/degrees")

# Create new Veg-X document with site characteristics
x = addSiteCharacteristics(newVegX(), moki_site, mapping = sitemapping,
                           measurementMethods = list(slope = slopeDeg, aspect = aspectDeg))

# Inspect the result
showElementTable(x, "plot")

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other add functions: \code{\link{addAggregateOrganismObservations}},
  \code{\link{addIndividualOrganismObservations}},
  \code{\link{addPlotGeometries}},
  \code{\link{addPlotLocations}},
  \code{\link{addPlotObservations}},
  \code{\link{addSiteObservations}},
  \code{\link{addStratumObservations}},
  \code{\link{addSurfaceCoverObservations}},
  \code{\link{addTaxonBySiteData}}
}
