% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStratumObservations.R
\name{addStratumObservations}
\alias{addStratumObservations}
\title{Add stratum observation records}
\usage{
addStratumObservations(target, x, mapping, methods, stratumDefinition,
  missing.values = c(NA, ""), verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one aggregate taxon observation. Columns can be varied.}

\item{mapping}{A list with element names 'plotName', 'obsStartDate', and 'stratumName' used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Additional optional mappings are: 'obsEndDate', 'subPlotName', 'lowerLimitMeasurement', 'lowerLimitMeasurement', and mappings to other stratum measurements.}

\item{methods}{A list measurement methods for stratum measurements (an object of class \code{\linkS4class{VegXMethod}}).}

\item{stratumDefinition}{An object of class \code{\linkS4class{VegXStrata}} indicating the definition of strata.}

\item{missing.values}{A character vector of values that should be considered as missing observations/measurements.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds stratum observation records to a VegX object from a data table
}
\details{
Missing plotName or obsStartDate values are interpreted as if the previous non-missing value has to be used to define plot observation.
Missing subPlotName values are interpreted in that observation refers to the parent plotName.
Missing measurements are simply not added to the Veg-X document.
}
\examples{
# Load source data
data(mokihinui)

# Define mapping
mapping = list(plotName = "Plot", obsStartDate = "PlotObsStartDate", stratumName = "Tier",
               lowerLimitMeasurement = "TierLower", upperLimitMeasurement = "TierUpper",
               cover = "CoverClass")


# Define cover scale
coverscale = defineOrdinalScaleMethod(name = "Recce cover scale",
                   description = "Recce recording method by Hurst/Allen",
                   subject = "plant cover",
                   citation = "Hurst, JM and Allen, RB. (2007) The Recce method for describing New Zealand vegetation – Field protocols. Landcare Research, Lincoln.",
                   codes = c("P","1","2","3", "4", "5", "6"),
                   quantifiableCodes = c("1","2","3", "4", "5", "6"),
                   breaks = c(0, 1, 5, 25, 50, 75, 100),
                   midPoints = c(0.05, 0.5, 15, 37.5, 62.5, 87.5),
                   definitions = c("Presence", "<1\%", "1-5\%","6-25\%", "26-50\%", "51-75\%", "76-100\%"))

# Define height measurement methods
heightMethod = predefinedMeasurementMethod("Stratum height/m")

# Define strata
strataDef = defineMixedStrata(name = "Recce strata",
                              description = "Standard Recce stratum definition",
                              citation = "Hurst, JM and Allen, RB. (2007) The Recce method for describing New Zealand vegetation – Field protocols. Landcare Research, Lincoln.",
                              heightStrataBreaks = c(0, 0.3,2.0,5, 12, 25, 50),
                              heightStrataNames = paste0("Tier ",1:6),
                              categoryStrataNames = "Tier 7",
                              categoryStrataDefinition = "Epiphytes")

# Create new Veg-X document with stratum observations
x = addStratumObservations(newVegX(), tier, mapping = mapping,
                        methods = list(lowerLimitMeasurement = heightMethod,
                                       upperLimitMeasurement = heightMethod,
                                       cover=coverscale),
                        stratumDefinition = strataDef)

# Examine results
summary(x)
head(showElementTable(x, "stratumObservation"))

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other add functions: \code{\link{addAggregateOrganismObservations}},
  \code{\link{addIndividualOrganismObservations}},
  \code{\link{addPlotLocations}},
  \code{\link{addPlotObservations}},
  \code{\link{addSiteCharacteristics}},
  \code{\link{addSiteObservations}},
  \code{\link{addTaxonBySiteData}}
}
