% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTaxonBySiteData.R
\name{addTaxonBySiteData}
\alias{addTaxonBySiteData}
\title{Add a taxon by site table}
\usage{
addTaxonBySiteData(target, x, abundanceMethod, obsDates = Sys.Date(),
  isTaxon = TRUE, absence.values = c(NA, 0), verbose = TRUE)
}
\arguments{
\item{target}{The object of class \code{\linkS4class{VegX}} to be modified.}

\item{x}{A taxon by site (plot) table (with plots in rows and taxa in columns) and with plant abundance values.}

\item{abundanceMethod}{A measurement method for aggregate plant abundance (an object of class \code{\linkS4class{VegXMethodDefinition}}).}

\item{obsDates}{A \code{\link{Date}} or a vector of \code{\link{Date}} objects with plot observation dates.}

\item{isTaxon}{A boolean flag (or vector) to indicate which organism names are taxa.}

\item{absence.values}{A vector of values to be interpreted as missing plant information.}

\item{verbose}{A flag to indicate console output of the data integration process.}
}
\value{
A modified object of class \code{\linkS4class{VegX}}
}
\description{
Adds the aggregate organism observations of a taxon by site table to an existing VegX object.
}
\seealso{
Other add functions: \code{\link{addAggregateOrganismObservations}},
  \code{\link{addCommunityObservations}},
  \code{\link{addIndividualOrganismObservations}},
  \code{\link{addPlotGeometries}},
  \code{\link{addPlotLocations}},
  \code{\link{addPlotObservations}},
  \code{\link{addSiteCharacteristics}},
  \code{\link{addSiteObservations}},
  \code{\link{addStratumObservations}},
  \code{\link{addSurfaceCoverObservations}}
}
