% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTaxonConcepts.R
\name{addTaxonConcepts}
\alias{addTaxonConcepts}
\title{Add taxon concepts}
\usage{
addTaxonConcepts(target, citationStringAll = "", x = NULL,
  mapping = list(), date.format = "\%Y-\%m-\%d", missing.values = c(NA,
  "0", ""), verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{citationStringAll}{A string with the bibliographic citation to be applied to all organism identities of the VegX object (using the original organism names as taxon names),
or to all original organism names listed in \code{x}.}

\item{x}{A data frame where each row corresponds to one organism identity, given by a column that can be mapped to \code{originalOrganismName}.}

\item{mapping}{A named list whose elements are strings that correspond to column names in \code{x}. Names of the list should be:
\itemize{
  \item{\code{originalOrganismName} - A string with the original name given by the author of the data set (required).}
  \item{\code{taxonName} - A string with the taxon name forming the taxon concept, if different from \code{originalOrganismName} (optional).}
  \item{\code{citationString} - A string with the bibliographic citation forming the taxon concept (optional).}
  \item{\code{assertionDate} - Date of taxon concept assertion (see \code{date.format}) (optional).}
  \item{\code{assertionParty} - Name of the party that undertook taxon concept assertion (optional).}
}}

\item{date.format}{A character string specifying the input format of dates (see \code{\link{as.Date}}).}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Sets or resets taxon concepts associated with organism identities
}
\examples{

  # TO BE DONE
}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other add functions: \code{\link{addAggregateOrganismObservations}},
  \code{\link{addIndividualOrganismObservations}},
  \code{\link{addPlotGeometries}},
  \code{\link{addPlotLocations}},
  \code{\link{addPlotObservations}},
  \code{\link{addSiteCharacteristics}},
  \code{\link{addSiteObservations}},
  \code{\link{addStratumObservations}},
  \code{\link{addSurfaceCoverObservations}},
  \code{\link{addTaxonBySiteData}}
}
