% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTaxonObservationRecords.R
\name{addTaxonObservationRecords}
\alias{addTaxonObservationRecords}
\title{Add taxon observation records}
\usage{
addTaxonObservationRecords(target, x, projectTitle, mapping, abundanceMethod,
  stratumDefinition = NULL, missing.values = c(NA, 0), verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one aggregated taxon observation. Columns can be varied.}

\item{projectTitle}{A character string to identify the project title, which can be the same as one of the currently defined in \code{target}.}

\item{mapping}{A list with element names 'plotName', 'obsStartDate', 'taxonAuthorName' and 'value', used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Additional optional mappings are: 'subPlotName', 'obsEndDate' and 'stratumName'.}

\item{abundanceMethod}{Measurement method for aggregated plant abundance (an object of class \code{\linkS4class{VegXMethod}}).}

\item{stratumDefinition}{An object of class \code{\linkS4class{VegXStrata}} indicating the definition of strata.}

\item{missing.values}{A vector of abundance values that should be considered as missing observations.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds aggregated taxon observation records to a VegX object from a data table
using a mapping to identify columns: plot, observation date, stratum, taxon name and value.
}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
\code{\link{addTreeObservationRecords}}.
}
