% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTaxonObservations.R
\name{addTaxonObservations}
\alias{addTaxonObservations}
\title{Add taxon observation records}
\usage{
addTaxonObservations(target, x, projectTitle, mapping, abundanceMethod,
  stratumDefinition = NULL, missing.values = c(NA, "0", ""),
  verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one aggregate taxon observation. Columns can be varied.}

\item{projectTitle}{A character string to identify the project title, which can be the same as one of the currently defined in \code{target}.}

\item{mapping}{A list with element names 'plotName', 'obsStartDate', 'authorTaxonName' and 'value', used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Additional optional mappings are: 'subPlotName', 'obsEndDate' and 'stratumName'.}

\item{abundanceMethod}{Measurement method for aggregate plant abundance (an object of class \code{\linkS4class{VegXMethod}}).}

\item{stratumDefinition}{An object of class \code{\linkS4class{VegXStrata}} indicating the definition of strata.}

\item{missing.values}{A character vector of values that should be considered as missing observations/measurements.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds aggregate taxon observation records to a VegX object from a data table
using a mapping to identify columns: plot, observation date, stratum, taxon name and value.
}
\examples{
data(mokihinui)

# Create new Veg-X document
target = newVegX()

# Define mapping
mapping = list(plotName = "Plot", obsStartDate = "obsDate", authorTaxonName = "PreferredSpeciesName",
              stratumName = "Tier", value = "Category")

# Define abundance scale
scale = definePlantCoverScale(name = "Recce cover scale", description = "Recce recording method by Allen",
                         citation = "Hurst, JM and Allen, RB. (2007) The Recce method for describing New Zealand vegetation – Field protocols. Landcare Research, Lincoln.",
                         breaks = c(0, 0.1, 1, 5, 25, 50, 75, 100),
                         midPoints = c(0.01, 0.05, 0.5, 15, 37.5, 62.5, 87.5),
                         values = c("P","1","2","3", "4", "5", "6"))

# Define strata
strataDef = defineMixedStrata(name = "Recce strata",
                              description = "Standard Recce stratum definition",
                              citation = "Hurst, JM and Allen, RB. (2007) The Recce method for describing New Zealand vegetation – Field protocols. Landcare Research, Lincoln.",
                              heightStrataBreaks = c(0, 0.3,2.0,5, 12, 25, 50),
                              heightStrataNames = paste0("Tier ",1:6),
                              categoryStrataNames = "Tier 7",
                              categoryStrataDefinition = "Epiphytes and vines")

# Mapping process
x = addTaxonObservations(target, tcv, "Mokihinui",
                        mapping = mapping,
                        abundanceMethod = scale,
                        stratumDefinition = strataDef)

summary(x)

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other add functions: \code{\link{addPlotLocations}},
  \code{\link{addSiteCharacteristics}},
  \code{\link{addSiteObservations}},
  \code{\link{addTaxonBySiteData}},
  \code{\link{addTreeObservations}}
}
