% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCoverScale.R
\name{defineCoverScale}
\alias{defineCoverScale}
\title{Cover scale definition}
\usage{
defineCoverScale(name = "Braun-Blanquet",
  description = "Five-level Braun-Blanquet cover scale", breaks = c(0, 5,
  25, 50, 75, 100), midPoints = c(2.5, 17.5, 37.5, 62.5, 87.5),
  values = as.character(1:5))
}
\arguments{
\item{name}{String with the name of the cover scale}

\item{description}{String describing the cover scale}

\item{breaks}{A vector of cover break points}

\item{midPoints}{A vector of cover class midpoints}

\item{values}{A character vector of class codes}
}
\value{
an object of class \code{\linkS4class{VegXMethod}}
}
\description{
Defines a cover scale for aggregated organism observations or stratum observations
}
\examples{
# Define abundance scale
scale = defineCoverScale(name = "Standard Recce (Allen)", description = "Recce recording method by Allen",
                        breaks = c(0, 0.1, 1, 5, 25, 50, 75, 100), midPoints = c(0.01, 0.05, 0.5, 15, 37.5, 62.5, 87.5),
                        values = c("P","1","2","3", "4", "5", "6"))

}
