% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineStrata.R
\name{defineHeightStrata}
\alias{defineHeightStrata}
\title{Stratum definition by height}
\usage{
defineHeightStrata(name = "Strata by height",
  description = "Vegetation strata defined by height in m", citation = "",
  heightBreaks = c(0, 1, 3, 5), stratumNames = c("s1", "s2", "s3"),
  heightUnit = "m")
}
\arguments{
\item{name}{A string to identify the stratum definition.}

\item{description}{A string describing how strata are defined.}

\item{citation}{A string with the bibliographic reference for the method.}

\item{heightBreaks}{A numeric vector with height limits between strata (of length equal to the number of strata plus one).}

\item{stratumNames}{A numeric vector of stratum codes (of length equal to the number of strata).}

\item{heightUnit}{A string to identify height units.}
}
\value{
an object of class \code{\linkS4class{VegXStrata}}
}
\description{
Height-based strata definition
}
