% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineOrdinalScale.R
\name{defineOrdinalScale}
\alias{defineOrdinalScale}
\title{Ordinal scale definition}
\usage{
defineOrdinalScale(name, description, subject, values, citation = "",
  breaks = NULL, midPoints = NULL)
}
\arguments{
\item{name}{String with the name of the ordinal scale}

\item{description}{String describing the ordinal scale}

\item{subject}{A string to identify the subject}

\item{values}{A character vector of class codes}

\item{citation}{A string with the bibliographic reference for the method.}

\item{breaks}{A vector of break points (for ordinal scales with class limits)}

\item{midPoints}{A vector of class midpoints (for ordinal scales that can be translated into quantitative values)}
}
\value{
an object of class \code{\linkS4class{VegXMethod}}
}
\description{
Defines an oridnal scale for a given subject (measurement)
}
