% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definePlantCoverScale.R
\name{definePlantCoverScale}
\alias{definePlantCoverScale}
\title{Cover scale definition}
\usage{
definePlantCoverScale(name = "Braun-Blanquet",
  description = "Five-level Braun-Blanquet cover scale", citation = "",
  breaks = c(0, 5, 25, 50, 75, 100), midPoints = c(2.5, 17.5, 37.5, 62.5,
  87.5), values = as.character(1:5))
}
\arguments{
\item{name}{String with the name of the cover scale}

\item{description}{String describing the cover scale}

\item{citation}{A string with the bibliographic reference for the method.}

\item{breaks}{A vector of cover break points}

\item{midPoints}{A vector of cover class midpoints}

\item{values}{A character vector of class codes}
}
\value{
an object of class \code{\linkS4class{VegXMethod}}
}
\description{
Defines a cover scale for aggregated organism observations or stratum observations
}
\examples{
scale = definePlantCoverScale(name = "Recce cover scale", description = "Recce recording method by Allen",
                              citation = "Hurst, JM and Allen, RB. (2007) The Recce method for describing New Zealand vegetation – Field protocols. Landcare Research, Lincoln.",
                              breaks = c(0, 0.1, 1, 5, 25, 50, 75, 100), midPoints = c(0.01, 0.05, 0.5, 15, 37.5, 62.5, 87.5),
                              values = c("P","1","2","3", "4", "5", "6"))

}
