% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineQuantitativeScaleMethod.R
\name{defineQuantitativeScaleMethod}
\alias{defineQuantitativeScaleMethod}
\title{Quantitative scale method definition}
\usage{
defineQuantitativeScaleMethod(name, description, subject, unit, citation = "",
  lowerLimit = -Inf, upperLimit = Inf)
}
\arguments{
\item{name}{String with the name of the method.}

\item{description}{String describing the method.}

\item{subject}{A string to identify the subject.}

\item{unit}{A character describing measurement units.}

\item{citation}{A string with the bibliographic reference for the method.}

\item{lowerLimit}{The lower limit of the quantitative scale, if defined.}

\item{upperLimit}{The upper limit of the quantitative scale, if defined.}
}
\value{
an object of class \code{\linkS4class{VegXMethod}}
}
\description{
A general function to define a method for measurements in a quantitative scale of a given subject
}
\examples{

defineQuantitativeScaleMethod(name = "Percent cover",
                              description = "Quantitative plant percent cover",
                              unit = "\%",
                              subject = "plant cover",
                              lowerLimit = 0, upperLimit = 100)

}
\seealso{
Other define measurement functions: \code{\link{defineOrdinalScaleMethod}},
  \code{\link{defineQualitativeScaleMethod}},
  \code{\link{predefinedMeasurementMethod}}
}
