% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineStrataByHeight.R
\name{defineStratumCategories}
\alias{defineStratumCategories}
\title{Strata definition by simple category}
\usage{
defineStratumCategories(name = "Strata by categories",
  description = "Vegetation categorical strata", citation = "",
  stratumNames = c("s1", "s2", "s3"))
}
\arguments{
\item{name}{A string to identify the stratum definition.}

\item{description}{A string describing how strata are defined.}

\item{citation}{A string with the bibliographic reference for the method.}

\item{stratumNames}{A numeric vector of stratum codes (of length equal to the number of strata).}
}
\value{
an object of class \code{\linkS4class{VegXStrata}}
}
\description{
Strata definition by simple category
}
