% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineSurfaceTypes.R
\name{defineSurfaceTypes}
\alias{defineSurfaceTypes}
\title{Surface type definition}
\usage{
defineSurfaceTypes(
  name,
  description,
  surfaceNames,
  definitions = NULL,
  citationString = "",
  DOI = ""
)
}
\arguments{
\item{name}{A string to identify the surface type definition.}

\item{description}{A string describing how surface types are defined.}

\item{surfaceNames}{A character vector of names for surface types.}

\item{definitions}{A character vector of definition of surface types.}

\item{citationString}{A string with the bibliographic reference for the method.}

\item{DOI}{A string with the DOI of a resource describing the method.}
}
\value{
an object of class \code{\linkS4class{VegXSurfaceTypeDefinition}}
}
\description{
Surface type definition
}
\examples{

defineSurfaceTypes(name = "Default Surface types",
                   description = "Five simple surface categories",
                   surfaceNames = c("Vegetation", "Moss", "Litter", "Exposed Soil", "Rock"))

}
