% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadVegXschema.R
\name{downloadVegXschema}
\alias{downloadVegXschema}
\title{Donwload VegX schema}
\usage{
downloadVegXschema(folder = getwd())
}
\arguments{
\item{folder}{The latest schema will be downloaded to this folder}
}
\description{
Downloads VegX schema, storing files in the indicated folder
}
\examples{
\dontrun{
#Downloads schema to the current folder
downloadVegXschema()
}
}
\author{
Sebastian Schmidtlein
}
