% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillPartyInformation.R
\name{fillPartyInformation}
\alias{fillPartyInformation}
\title{Fills the information for a given party}
\usage{
fillPartyInformation(target, name, type = "individual", address = "",
  phone = "", electronicMailAddress = "", onlineURL = "")
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified.}

\item{name}{A string with the full name of the person/organization/position being described.}

\item{address}{A string that identifies the party type: "individual", "organization" or "position".}

\item{phone}{A string summarizing the aims and findings of the project.}

\item{electronicMailAddress}{A string with information about funding agencies.}

\item{onlineURL}{A string describing the study area succinctly.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Fills the information for a given party
}
\examples{

x = fillPartyInformation(newVegX(), name = "Susan K. Wiser")

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other fill functions: \code{\link{fillProjectInformation}}
}
