% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillProjectInformation.R
\name{fillProjectInformation}
\alias{fillProjectInformation}
\title{Fills the information for a given research project}
\usage{
fillProjectInformation(target, title, personnel = list(), abstract = "",
  funding = "", studyAreaDescription = "", designDescription = "",
  verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified.}

\item{title}{A string with the title of the project. It can match one of the existing project or a new one will be defined.}

\item{personnel}{A named list where element names are roles and values are party names (e.g. \code{list(contributor = "John Smith")}).}

\item{abstract}{A string summarizing the aims and findings of the project.}

\item{funding}{A string with information about funding agencies.}

\item{studyAreaDescription}{A string describing the study area succinctly.}

\item{designDescription}{A string describing the overall plot placement design.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Fills the information for a given research project
}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
