% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationDatasets.r
\docType{data}
\name{mtfyffe}
\alias{mtfyffe}
\title{Mt Fyffe example dataset for Veg-X}
\description{
The data set concerns observations made on 5 plots with their subplots.
The data set includes six data frames:
}
\details{
\itemize{
 \item mtfyffe_loc Location of vegetation plots (10 records x 25 variables)
 \item mtfyffe_site Site characteristics of plots/subplots (325 records x 29 variables)
 \item mtfyffe_counts Observations of taxon categories, grouped by vegetation strata (543 records x 30 variables)
 \item mtfyffe_dia Observations made on individual trees (880 records x 31 variables)
 \item mtfyffe_disturbance Observations of disturbance events on plots (9 records x 10 variables)
 \item mtfyffe_groundcover Observations of ground cover on plots (50 records x 12 variables)
}
}
\references{
Wiser, S.K., P.J. Bellingham and L.E. Burrows.  2001.  Managing biodiversity information:  development of the National Vegetation Survey Databank.  New Zealand Journal of Ecology 25(2): 1-17.
}
\author{
NVS database. Landcare Research (New Zealand) \email{nvs@landcareresearch.co.nz}
}
\keyword{data}
