% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newVegX.R
\name{newVegX}
\alias{newVegX}
\title{New VegX document}
\usage{
newVegX(projectTitle)
}
\arguments{
\item{projectTitle}{a character string to identify the project title}
}
\value{
an object of class \code{\linkS4class{VegX}}
}
\description{
Creates a new (empty) VegX document
}
