% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationMethods.r
\docType{data}
\name{ordinal_methods}
\alias{ordinal_methods}
\title{List of predefined ordinal methods}
\description{
This list contains the definition of predefined ordinal methods
  available for Veg-X.

The main elements contained in this list of methods are:
\itemize{
 \item name The name of the method
 \item description Short description of the method
 \item subject The description of an attribute class for comparative
 purposes.
 \item citationString The bibliographic reference containing the method
 description (if available)
 \item code The ordinal codes of the ordinal variable
 \item definition The definitions of the codes of the ordinal variable
 \item lowerLimit, upperLimit and midPoint The limits and midpoint of the
 ordinal codes
 
}
}
\examples{

# To explore available methods
names(ordinal_methods)

}
\seealso{
\code{\link{defineOrdinalScaleMethod}}
}
\author{
Veg-X team.
}
\keyword{data}
