% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareInputData.R
\name{prepareInputData}
\alias{prepareInputData}
\title{Prepare Data to the VegX Format}
\usage{
prepareInputData(
  plot.info = NULL,
  plot.obs = NULL,
  site.data = NULL,
  surface.data = NULL,
  community.data = NULL,
  species.data = NULL,
  individual.data = NULL,
  project.info = NULL,
  people.info = NULL,
  citation.info = NULL,
  mapping = NULL,
  project.title = NULL,
  miss.fields = TRUE,
  fixed.date = Sys.Date(),
  spp.names = c("organismNameOriginal", "organismNameMorpho", "organismName"),
  summarise.data = TRUE,
  profiles = NULL,
  return.unmapped = FALSE
)
}
\arguments{
\item{plot.info}{a data frame. General information of the plots or invariant
inventory site characteristics (e.g. plot or site codes, names and
identifiers, methods, location (place names), geographical coordinates,
elevation, aspect, slope). If the site characteristic can vary between
census, it should be provided as `site.data`.}

\item{plot.obs}{a data frame. Observations or census information related to
the plots or inventories (e.g. census dates, data generator, citation,
licenses, taxonomic quality checks)}

\item{site.data}{a data frame. Observations or measurements associated with
the inventory site that may vary in time (soil and climate conditions,
legal protection, land use, vegetation type)}

\item{surface.data}{a data frame. Surface cover measurements associated with
the plot at a given time period.}

\item{community.data}{a data frame. Observations or measurements associated
with the entire vegetation, i.e. observations at the stand or community
level (e.g. density, biomass, species diversity, canopy cover, vegetation
height, successional status)}

\item{species.data}{a data frame. The aggregated observations per species
(e.g. species name, count of individuals, total biomass, frequency)}

\item{individual.data}{a data frame. The observations per plant individual
(e.g. species name, field tag, diameter, height, coordinates within the
plot)}

\item{project.info}{a data frame. The information about the project that
generated/compiled the vegetation data (e.g. abstract, funding, citation
and short descriptions of the studied area and the typical project design)}

\item{people.info}{a data frame. The information about the people involved in
the project that generated/compiled the vegetation data (e.g. full names,
roles, addresses, email)}

\item{citation.info}{a data frame. The information about the literature
associated with the vegetation data (e.g. citation ID, authors, year,
title, etc.)}

\item{mapping}{a named list with no default. Usually the output of the VegX
function function `prepareMapping()`.}

\item{project.title}{a character with the project name/title. Only needed if
the user wants to use a different name than the one provided in the
argument `mapping`.}

\item{miss.fields}{logical. Should any missing fields be auto-completed in
the field mapping from matching the names of the columns in the input data?
Defaults to TRUE.}

\item{fixed.date}{date. A fixed date to be added for projects or plots
without census date information. Defaults to the current system data (i.e.
`Sys.Date()`). For dates other than the default, preferably use the
following format: "YYYY-MM-DD".}

\item{spp.names}{character vector. The priority order for defining the
working organism name for building the VegX object and for calculating
species diversity indices. Defaults to the following order:
'organismNameMorpho', 'organismNameOriginal' and 'organismName'.}

\item{summarise.data}{logical. Should individual or species plot data be
aggregated at higher levels of information (i.e. species and community
levels)? Defaults to TRUE.}

\item{profiles}{numerical vector. The plant diameter class limits to generate
different data profiles for the individual data when aggregated at the
species data.}

\item{return.unmapped}{logical. Should the data contained in fileds not
present in the map provided be saved with the output. Defaults to FALSE.}
}
\value{
A named list with the names of the fields essential for VegX
  (`essential.info`), the updated mapping (`mapping`), the organized data
  present in the mapping `data` and, if chosen, the as-provided data
  not present in the mapping (`unmapped.data`).
}
\description{
This function prepares, check and organize the available
  information to ease the creation of VegX objects.
}
\details{
Some fields, such as the the plot ID/name, sampling date and
  organism names, are quite important for building the VegX objects. So, if
  missing and if possible, they are created and added to the data. In
  addition, if data at species and/or individual-level are provided, plot
  ID/name need to be present across the provided tables.
  
  If not data is provided with the species and/or individual-level, the
  function assumes that the data is single-census and add the sampling dates
  as provided in the plot information/observation tables. If no sampling date
  is provided, the function assumes by default the date when the data was
  prepared. Users can provide a different date using the argument
  `fixed.date`, preferably in the "YYYY-MM-DD" format.
  
  VegX only accepts complete organism names (e.g. Genus epiteth Author). If
  organism name information is splitted into different columns (e.g. genus,
  epiteth, infra-epiteth, author), those columns are used to generate full
  organism names, if they were also provided in the mapping. This operation
  is performed for original names, morpho-species or accepted names.
  
  If a given field containing a quantitative or qualitative
  measurement/classification is present in the mapping and it has not an
  associate method description, the field is dropped from the data and the
  mapping with a warning. The field will still be stored in the output under
  the element `unmapped.data` if the argument `return.unmapped` is set to
  TRUE.
  
  If species and/or individual-level data is provided and mapped, the
  function by default returns the aggregated data at the levels possible 
  (i.e. species and/or community levels) and it include the corresponding
  methods in the mapping. When individual-level data is provided, the
  aggregated data at the species-level tipically contains the count of number
  of individuals and the sum of their basal area per organism name. The
  organism name selected to obtain these metrics is the first one from the
  `spp.names` vectors which is also present in the data. If only
  species-level data is provided, then only metrics at the stand/community
  level are obtained. Currently those metrics are: the number of individuals
  in the sample, the number of individuals, basal area and above-ground
  biomass per hectare, the number of species in the sample (i.e. richness),
  and the Shannon, Simpson and Pielou indices. These calculations can be
  avoided by setting the argument `summarise.data` to FALSE.
}
\seealso{
Other high-level functions: 
\code{\link{buildVegX}()},
\code{\link{prepareMapping}()}
}
\concept{high-level functions}
