% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMapping.R
\name{prepareMapping}
\alias{prepareMapping}
\title{Map Field Names}
\usage{
prepareMapping(
  project = NULL,
  equivalencies = NULL,
  user.map = NULL,
  user.column = "YourFieldNames"
)
}
\arguments{
\item{project}{a character. The acronym or name of a network or project to
obtain pre-defined field maps (see details).}

\item{equivalencies}{a data frame. A user-defined map of a given dataset to
the to standard field and method names (see details).}

\item{user.map}{a named list. The user-defined mapping of the dataset to
standard field and method names (see details).}

\item{user.column}{a character. The name of the column in
\code{equivalencies} containing the mapping. Defaults to 'YourFieldNames'.}
}
\value{
A named list, with named vectors as elements.
}
\description{
Provides different options to map equivalencies of the fields and the methods
used in a given dataset to standard field and method names.
}
\details{
Users can map their dataset to the standard field names and methods
  in three different ways:
  
  \itemize{
  \item{Option 1: provide to the argument \code{project} the acronym of
  one of the network, groups or databases of vegetation data with available
  pre-defined mapping to the field and method standards. The currently
  available options are available in the internal object
  \code{\link{supporting_info}}.}
   
  \item{Option 2: provide to the argument \code{equivalencies} a data
  frame containing a user-defined mapping of the dataset. Users can download
  and complete this data frame using the model and explanations available
  \href{INCLUDELINKHERE}{here}.}

  \item{Option 3: provide to the argument \code{user.map} a named list
  containing a user-defined mapping of the dataset. This named list must
  follow the standard names which are in the internal object
  \code{\link{supporting_info}}. More information on the 
  meaning of each field are available \href{INCLUDELINKHERE}{here}.}
  }
  
  If the user chose for Option 2, than the use should also provide the name
  of the column containing the mapping. If the name of the column is not
  present in the input data frame, the function assumes the 4th column to be
  the one containing the mapping. If even the 4th column is not different
  from the predefined fields of the internal standard, then the function uses
  the next column which is not a predefined field as a final try.
}
\examples{
# Example of the Option 1
 mapping <- prepareMapping(project = "NVS") 

# Example of the Option 3
 my_map <- list(plot.info = c(plotName = "Plot",  
                             samplingEffort = "PlotArea", 
                             samplingEffortMethod = "Plot area/m2",
                             plotLongitude = "Longitude", 
                             plotLatitude = "Latitude",
                             elevationSite = "Altitude",
                             elevationSiteMethod = "Elevation/m",
                             censusDateStart = "PlotObsStartDate", 
                             censusDateStop = "PlotObsStopDate"),
              individual.data = c(plotNameSample = "Plot",
                                  obsStartDateSample = "PlotObsStartDate",
                                  fieldTag = "Identifier",
                                  organismNameOriginal = "NVSSpeciesName",
                                  diameter = "Diameter",
                                  diameterMethod = "DBH/cm"))
 
 mapping <- prepareMapping(user.map = my_map)
 
}
\seealso{
Other high-level functions: 
\code{\link{buildVegX}()},
\code{\link{prepareInputData}()}
}
\concept{high-level functions}
