% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationMethods.r
\docType{data}
\name{qualitative_methods}
\alias{qualitative_methods}
\title{List of predefined qualitative methods}
\description{
This list contains the definition of predefined qualitative
  methods available for Veg-X.

The main elements contained in this list of methods are:
\itemize{
 \item name The name of the method
 \item description Short description of the method
 \item subject The description of an attribute class for comparative purposes.
 \item citationString The bibliographic reference containing the method
 description (if available)
 \item code The class codes of the qualitative variable
 \item definition The definitions of the class codes of the qualitative variable
}
}
\examples{

# To explore available methods
names(qualitative_methods)

}
\seealso{
\code{\link{defineQualitativeScaleMethod}}
}
\author{
Veg-X team.
}
\keyword{data}
