% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVegXML.R
\name{readVegXML}
\alias{readVegXML}
\title{Reads a Veg-X XML file}
\usage{
readVegXML(file)
}
\arguments{
\item{file}{the filename of an XML file following the Veg-X standard (ver. 2.0)}
}
\value{
An object of class \code{\linkS4class{VegX}}
}
\description{
Reads a Veg-X XML file
}
