% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPreferredTaxonNomenclature.R
\name{setPreferredTaxonNomenclature}
\alias{setPreferredTaxonNomenclature}
\title{Sets preferred taxon nomenclature}
\usage{
setPreferredTaxonNomenclature(target, x, mapping,
  date.format = "\%Y-\%m-\%d", missing.values = c(NA, ""),
  verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to a different value of 'originalOrganismName'.}

\item{mapping}{A named list whose elements are strings that correspond to column names in \code{x}. Names of the list should be:
\itemize{
  \item{\code{originalOrganismName} - A string with the original name given by the author of the data set (required).}
  \item{\code{preferredTaxonName} - A string with the preferred taxon name forming the taxon (required).}
  \item{\code{interpretationSource} - A string describing the source for the last nomenclature interpretation applied to this organism identity (i.e. the Plant List). (optional).}
  \item{\code{interpretationCitation} - A string of the publication where nomenclature interpretation is explained (optional).}
  \item{\code{interpretationDate} - Date of taxon nomenclature interpretation (see \code{date.format}) (optional).}
  \item{\code{interpretationParty} - Name of the party that undertook nomenclature interpretation (optional).}
}}

\item{date.format}{A character string specifying the input format of dates (see \code{\link{as.Date}}).}

\item{missing.values}{A character vector of values that should be considered as missing data (see details).}

\item{verbose}{A boolean flag to indicate console output of the nomenclatural change process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Sets (or resets) a preferred taxon nomenclature to the organism identities of a data set using a lookup table
}
\examples{
# Load source data
data(mokihinui)

# Create new Veg-X document with aggregate organism observations
mapping = list(plotName = "Plot", obsStartDate = "PlotObsStartDate",
               taxonName = "NVSSpeciesName",
               stratumName = "Tier", cover = "Category")
coverscale = defineOrdinalScaleMethod(name = "Recce cover scale",
                   description = "Recce recording method by Hurst/Allen",
                   subject = "plant cover",
                   citation = "Hurst, JM and Allen, RB. (2007) 
                       The Recce method for describing New Zealand vegetation – Field protocols. 
                       Landcare Research, Lincoln.",
                   codes = c("P","1","2","3", "4", "5", "6"),
                   quantifiableCodes = c("1","2","3", "4", "5", "6"),
                   breaks = c(0, 1, 5, 25, 50, 75, 100),
                   midPoints = c(0.05, 0.5, 15, 37.5, 62.5, 87.5),
                   definitions = c("Presence", "<1\%", "1-5\%","6-25\%", "26-50\%", 
                                   "51-75\%", "76-100\%"))
strataDef = defineMixedStrata(name = "Recce strata",
                   description = "Standard Recce stratum definition",
                   citation = "Hurst, JM and Allen, RB. (2007) 
                     The Recce method for describing New Zealand vegetation – Field protocols. 
                     Landcare Research, Lincoln.",
                   heightStrataBreaks = c(0, 0.3,2.0,5, 12, 25, 50),
                   heightStrataNames = paste0("Tier ",1:6),
                   categoryStrataNames = "Tier 7",
                   categoryStrataDefinition = "Epiphytes")
x = addAggregateOrganismObservations(newVegX(), moki_tcv,
                        mapping = mapping,
                        methods = c(cover=coverscale),
                        stratumDefinition = strataDef)

# Inspect the original organism identities
head(showElementTable(x, "organismIdentity"))

y = setPreferredTaxonNomenclature(x, moki_lookup,
           c(originalOrganismName = "NVSSpeciesName", preferredTaxonName = "PreferredSpeciesName"))

# Inspect the modified organism identities
head(showElementTable(y, "organismIdentity"))
}
\seealso{
Other organism identity functions: \code{\link{setOriginalIdentificationConcepts}}
}
\concept{organism identity functions}
