% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showElementTable.R
\name{showElementTable}
\alias{showElementTable}
\title{Coearces VegX elements into a data frame}
\usage{
showElementTable(x, element = "plot", includeIDs = FALSE)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{VegX}}}

\item{element}{The name of the main elements to be coerced: 'plot', 'plotObservation', 'taxonNameUsageConcept',
'stratum', 'stratumObservation', 'aggregateOrganismObservation', 'individualOrganism', 'individualOrganismObservation',
'siteObservation', 'method', 'attribute'.}

\item{includeIDs}{A boolean flag to indicate whether internal identifiers should be included in the output}
}
\value{
a data frame
}
\description{
Coerces part of the information of a Veg-X object into a data frame
}
\examples{
data(mokihinui)

# Create document 'x' with aggregate taxon observations
taxmapping = list(plotName = "Plot", obsStartDate = "obsDate", authorTaxonName = "PreferredSpeciesName",
              stratumName = "Tier", value = "Category")
scale = defineCoverScale(name = "Recce cover scale", description = "Recce recording method by Allen",
                         citation = "Hurst, JM and Allen, RB. (2007) The Recce method for describing New Zealand vegetation – Field protocols. Landcare Research, Lincoln.",
                         breaks = c(0, 0.1, 1, 5, 25, 50, 75, 100),
                         midPoints = c(0.01, 0.05, 0.5, 15, 37.5, 62.5, 87.5),
                         values = c("P","1","2","3", "4", "5", "6"))
strataDef = defineStrataByHeight(name = "Recce strata",
                                description = "Standard Recce stratum definition",
                                citation = "Hurst, JM and Allen, RB. (2007) The Recce method for describing New Zealand vegetation – Field protocols. Landcare Research, Lincoln.",
                                heightBreaks = c(0, 0.3,2.0,5, 12, 25,50, 100),
                                stratumNames = paste0("Tier ",1:7))
x = addTaxonObservations(newVegX(), tcv, "Mokihinui",
                        mapping = taxmapping,
                        abundanceMethod = scale,
                        stratumDefinition = strataDef)

# Summary
summary(x)

# show plot information
showElementTable(x, "plot")

# show plot observation information
showElementTable(x, "plotObservation")

# show methods and attributes
showElementTable(x, "method")
showElementTable(x, "attribute")

# show aggregate organism observations
showElementTable(x, "aggregateOrganismObservation")
}
