% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratumDefinitionByHeight.R
\name{stratumDefinitionByHeight}
\alias{stratumDefinitionByHeight}
\title{Stratum definition by height}
\usage{
stratumDefinitionByHeight(heightBreaks = c(0, 1, 3, 5),
  stratumNames = c("s1", "s2", "s3"), heightUnit = "m")
}
\arguments{
\item{heightBreaks}{A numeric vector with height limits between strata (of length equal to the number of strata plus one).}

\item{stratumNames}{A numeric vector of stratum codes (of length equal to the number of strata).}

\item{heightUnit}{A string to identify height units.}
}
\description{
User-defined strata based on height
}
