% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformOrdinalScale.R
\name{transformOrdinalScale}
\alias{transformOrdinalScale}
\title{Transform ordinal scale}
\usage{
transformOrdinalScale(target, method, newUnits, verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{method}{An integer (index) or a name of an ordinal scale method}

\item{verbose}{A boolean flag to indicate console output of the data transformation process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Replaces all the values made using an ordinal scale to their corresponding midpoint quantitative values
(e.g. cover scale values to percent cover).
}
