% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformQuantitativeScale.R
\name{transformQuantitativeScale}
\alias{transformQuantitativeScale}
\title{Transform quantitative scale}
\usage{
transformQuantitativeScale(target, method, newMethod, fun,
  replaceValues = FALSE, verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified.}

\item{method}{An integer (index) or a name of the original quantitative scale method.}

\item{newMethod}{An integer (index) or a name of a quantitative method existing in the initial object,
or an object of class \code{\linkS4class{VegXMethod}}.}

\item{fun}{A function used to transform numeric values.}

\item{replaceValues}{A boolean flag to indicate that values of the new scale should replace the old ones. 
For some measurements transformations will not be possible if replacement is not forced using this flag.}

\item{verbose}{A boolean flag to indicate console output of the data transformation process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Transforms all the values in a VegX object made using a quantitative scale into another quantitative scale
following a given transformation function.
}
\examples{
data(mokihinui)

}
\seealso{
Other transform functions: \code{\link{transformOrdinalScale}}
}
