% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformTaxonNomenclature.R
\name{transformTaxonNomenclature}
\alias{transformTaxonNomenclature}
\title{Applies a preferred taxon nomenclature to the organism identities of a data set}
\usage{
transformTaxonNomenclature(target, x, mapping, verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to a different value of 'originalOrganismName'.}

\item{mapping}{A named list with element names 'originalOrganismName', 'preferredTaxonName'}

\item{verbose}{A boolean flag to indicate console output of the nomenclatural change process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Applies a preferred taxon nomenclature to the organism identities of a data set
}
\examples{

}
\seealso{
Other transform functions: \code{\link{transformOrdinalScale}},
  \code{\link{transformQuantitativeScale}}
}
