% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateVegX.R
\name{validateVegX}
\alias{validateVegX}
\title{VegX XML file validation/check}
\usage{
validateVegX(file)
}
\arguments{
\item{file}{VegX xml file to be checked}
}
\description{
Validates an XML file against VegX schema and checks its internal logical consistency
}
\examples{
\dontrun{
# Load source data
data(mokihinui)

# Define mapping
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")

# Create a new Veg-X document with projects, plots and plot observations (no data)
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)

# Save file
writeVegXML(x, "foo.xml")

# Validate created XML file
validateVegX("foo.xml")

}
}
\author{
Sebastian Schmidtlein
}
