% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegXStrata.R
\docType{class}
\name{VegXStrata-class}
\alias{VegXStrata-class}
\title{S4 class for Veg-X strata definition}
\description{
S4 class for Veg-X strata definition
}
\section{Slots}{

\describe{
\item{\code{method}}{VegXMethod.}

\item{\code{strata}}{list.}
}}

