% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegXStrataDefinition.R
\docType{class}
\name{VegXStrataDefinition-class}
\alias{VegXStrataDefinition-class}
\title{S4 class for Veg-X strata definition}
\description{
S4 class for Veg-X strata definition
}
\section{Slots}{

\describe{
\item{\code{method}}{An object of class \code{\linkS4class{VegXMethod}}.}

\item{\code{strata}}{A list of vegetation strata.}
}}

\examples{
showClass("VegXStrataDefinition")

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
