% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegXSurfaceCoverDefinition.R
\docType{class}
\name{VegXSurfaceCoverDefinition-class}
\alias{VegXSurfaceCoverDefinition-class}
\title{S4 class for Veg-X surface cover definition}
\description{
S4 class for Veg-X surface cover definition
}
\section{Slots}{

\describe{
\item{\code{method}}{An object of class \code{\linkS4class{VegXMethod}}.}

\item{\code{surfaceCovers}}{A list of surface covers.}
}}

\examples{
showClass("VegXSurfaceCoverDefinition")

}
