% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegXSurfaceTypeDefinition.R
\docType{class}
\name{VegXSurfaceTypeDefinition-class}
\alias{VegXSurfaceTypeDefinition-class}
\title{S4 class for Veg-X surface type definition}
\description{
S4 class for Veg-X surface type definition
}
\section{Slots}{

\describe{
\item{\code{method}}{An object of class \code{\linkS4class{VegXMethodDefinition}}.}

\item{\code{surfaceTypes}}{A list of surface types.}
}}

\examples{
showClass("VegXSurfaceTypeDefinition")

}
