% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeVegXML.R
\name{writeVegXML}
\alias{writeVegXML}
\title{Write VegX XML file (ver 1.5.3)}
\usage{
writeVegXML(x, file)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{VegX}}}

\item{file}{the file name to be written}
}
\description{
Write VegX XML file (ver 1.5.3)
}
