% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAggregateOrganismObservations.R
\name{addAggregateOrganismObservations}
\alias{addAggregateOrganismObservations}
\title{Add aggregate organism observation records}
\usage{
addAggregateOrganismObservations(target, x, mapping, methods = list(),
  stratumDefinition = NULL, citationStringAll = NULL,
  date.format = "\%Y-\%m-\%d", missing.values = c(NA, "0", ""),
  verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one aggregate organism observation. Columns can be varied.}

\item{mapping}{A list with element names 'plotName', 'obsStartDate', 'organismName' or 'taxonName', and 'value', used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Additional optional mappings are: 'subPlotName', 'citationString' ,'stratumName', 'heightMeasurement' and mappings to other measurements (e.g. taxon abundance).}

\item{methods}{A list measurement methods for aggregate organism measurements (an object of class \code{\linkS4class{VegXMethodDefinition}}).}

\item{stratumDefinition}{An object of class \code{\linkS4class{VegXStrataDefinition}} indicating the definition of strata.}

\item{citationStringAll}{A string of a literature citation that explains the taxon concepts for all taxon names of the source data table.}

\item{date.format}{A character string specifying the input format of dates (see \code{\link{as.Date}}).}

\item{missing.values}{A character vector of values that should be considered as missing observations/measurements.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds aggregate organism observation records to a VegX object from a data table
}
\details{
Missing value policy:
\itemize{
  \item{Missing 'plotName' and 'obsStartDate' values are interpreted as if the previous non-missing value has to be used to define aggregate organism observation.}
  \item{Missing 'subPlotName' values are interpreted in that observation refers to the parent plotName.}
  \item{When 'stratumName' values are missing the aggregate organism observation is not assigned to any stratum.}
  \item{When both 'organismName' and 'taxonName' are missing (i.e. missing organism identity) the function generates an error.}
  \item{Missing aggregate organism measurements are not added to the Veg-X document.}
}
}
\examples{
# Load source data
data(mokihinui)

# Define mapping
mapping = list(plotName = "Plot", obsStartDate = "PlotObsStartDate",
               taxonName = "NVSSpeciesName",
               stratumName = "Tier", cover = "Category")

# Define abundance scale
coverscale = defineOrdinalScaleMethod(name = "Recce cover scale",
                   description = "Recce recording method by Hurst/Allen",
                   subject = "plant cover",
                   citation = "Hurst, JM and Allen, RB. (2007) The Recce method for describing New Zealand vegetation – Field protocols. Landcare Research, Lincoln.",
                   codes = c("P","1","2","3", "4", "5", "6"),
                   quantifiableCodes = c("1","2","3", "4", "5", "6"),
                   breaks = c(0, 1, 5, 25, 50, 75, 100),
                   midPoints = c(0.05, 0.5, 15, 37.5, 62.5, 87.5),
                   definitions = c("Presence", "<1\%", "1-5\%","6-25\%", "26-50\%", "51-75\%", "76-100\%"))

# Define strata
strataDef = defineMixedStrata(name = "Recce strata",
                              description = "Standard Recce stratum definition",
                              citation = "Hurst, JM and Allen, RB. (2007) The Recce method for describing New Zealand vegetation – Field protocols. Landcare Research, Lincoln.",
                              heightStrataBreaks = c(0, 0.3,2.0,5, 12, 25, 50),
                              heightStrataNames = paste0("Tier ",1:6),
                              categoryStrataNames = "Tier 7",
                              categoryStrataDefinition = "Epiphytes")

# Create new Veg-X document with aggregate organism observations
x = addAggregateOrganismObservations(newVegX(), moki_tcv,
                        mapping = mapping,
                        methods = c(cover=coverscale),
                        stratumDefinition = strataDef)

# Inspect the result
head(showElementTable(x, "aggregateOrganismObservation"))


# Example with individual counts
data(mtfyffe)
mapping = list(plotName = "Plot", subPlotName = "Subplot", obsStartDate = "PlotObsStartDate",
               taxonName = "NVSSpeciesName", stratumName = "Tier", counts = "Value")
countscale = predefinedMeasurementMethod("Plant counts")
strataDef = defineHeightStrata(name = "Standard seedling/sapling strata",
                               description = "Seedling/sapling stratum definition",
                               heightBreaks = c(0, 15, 45, 75, 105, 135, 200),
                               strataNames = as.character(1:6),
                               strataDefinitions = c("0-15 cm", "16-45 cm", "46-75 cm", "76-105 cm", "106-135 cm", "> 135 cm"))
x = addAggregateOrganismObservations(newVegX(), mtfyffe_counts, mapping,
                                     methods = c(counts=countscale),
                                     stratumDefinition = strataDef)
head(showElementTable(x, "aggregateOrganismObservation"))

# Example with frequency in transects
data(takitimu)
mapping = list(plotName = "Plot", subPlotName = "Subplot", obsStartDate = "PlotObsStartDate",
               taxonName = "NVSSpeciesName", freq = "Value")
freqscale = predefinedMeasurementMethod("Plant frequency/\%")
x = addAggregateOrganismObservations(newVegX(), taki_freq, mapping,
                                     methods = c(freq=freqscale))
head(showElementTable(x, "aggregateOrganismObservation"))

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other add functions: \code{\link{addIndividualOrganismObservations}},
  \code{\link{addPlotGeometries}},
  \code{\link{addPlotLocations}},
  \code{\link{addPlotObservations}},
  \code{\link{addSiteCharacteristics}},
  \code{\link{addSiteObservations}},
  \code{\link{addStratumObservations}},
  \code{\link{addSurfaceCoverObservations}},
  \code{\link{addTaxonBySiteData}}
}
