% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIndividualOrganismObservations.R
\name{addIndividualOrganismObservations}
\alias{addIndividualOrganismObservations}
\title{Add individual organism observation records}
\usage{
addIndividualOrganismObservations(target, x, projectTitle, mapping,
  methods = list(), stratumDefinition = NULL, missing.values = c(NA, "0",
  ""), verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one tree observation. Columns can be varied.}

\item{projectTitle}{A string to identify the project title, which can be the same as one of the currently defined in \code{target}.}

\item{mapping}{A list with element names 'plotName', 'obsStartDate', 'authorTaxonName' used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Additional optional mappings are: 'subPlotName', 'obsEndDate', 'individual', 'stratumName', 'diameterMeasurement', 'heightMeasurement', and names to identify additional specific measurements.}

\item{methods}{A named list of objects of class \code{\linkS4class{VegXMethod}} indicating the definition of 'diameterMeasurement', 'heightMeasurement' and any additional individual organism measurement defined in \code{mapping}.}

\item{stratumDefinition}{An object of class \code{\linkS4class{VegXStrata}} indicating the definition of strata.}

\item{missing.values}{A character vector of values that should be considered as missing observations/measurements.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds individual organism observation records to a VegX object from a data table.
}
\examples{
data(mokihinui)

# Create new Veg-X document
target = newVegX()

# Define mapping
mapping = list(plotName = "Plot", subPlotName = "Subplot", obsStartDate = "obsDate",
               authorTaxonName = "PreferredSpeciesName", diameterMeasurement = "Diameter")

# Define diameter measurement method
diamMeth = predefinedMeasurementMethod("DBH/cm")


# Mapping process
y = addIndividualOrganismObservations(target, dia, "Mokihinui",
                                      mapping = mapping,
                                      methods = c(diameterMeasurement = diamMeth))

# Inspect the result
summary(y)

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other add functions: \code{\link{addAggregateOrganismObservations}},
  \code{\link{addPlotLocations}},
  \code{\link{addSiteCharacteristics}},
  \code{\link{addSiteObservations}},
  \code{\link{addStratumObservations}},
  \code{\link{addTaxonBySiteData}}
}
