% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlotGeometries.R
\name{addPlotGeometries}
\alias{addPlotGeometries}
\title{Adds/replaces plot geometry information}
\usage{
addPlotGeometries(target, x, mapping, methods = list(),
  missing.values = c(NA, ""), verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified.}

\item{x}{A data frame where each row corresponds to one plot. Columns can be varied.}

\item{mapping}{A list with at least element name 'plotName', is used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Geometry variables that can be mapped are: 'area', 'shape', 'radius', 'length','width', 'bandWidth'.
Additional optional mappings are: 'subPlotName'.}

\item{methods}{A list measurement methods for plot geometry measurements (each being an object of class \code{\linkS4class{VegXMethodDefinition}}).}

\item{missing.values}{A character vector of values that should be considered as missing data.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds/replaces static plot geometry information (plot shape, dimensions, ...) to plot elements of a VegX object from a data table where rows are plots,
using a mapping to identify plot and subplot (optional). Additional mapping elements are used to map specific variables.
}
\details{
Missing value policy:
 \itemize{
    \item{Missing 'plotName' and 'shape' values are interpreted as if the previous non-missing value has to be used to define plot.}
    \item{Missing 'subPlotName' values are interpreted in that data refers to the parent plotName.}
    \item{Missing measurements (e.g. 'area', 'radius',...) are simply not added to the Veg-X document.}
 }
}
\examples{
# Load source data
data(mokihinui)

# Define location mapping
mapping = list(plotName = "Plot", subPlotName = "Subplot",
               area = "PlotArea", shape = "Shape",
               width = "PlotRectangleLength01", length = "PlotRectangleLength02")

# Define methods
areaMethod = predefinedMeasurementMethod("Plot area/m2")
dimensionMethod = predefinedMeasurementMethod("Plot dimension/m")

# Create new Veg-X document with plot locations
x = addPlotGeometries(newVegX(), moki_site, mapping,
                      list(area = areaMethod, width = dimensionMethod, length = dimensionMethod))

# Inspect results
showElementTable(x, "plot")

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other add functions: \code{\link{addAggregateOrganismObservations}},
  \code{\link{addIndividualOrganismObservations}},
  \code{\link{addPlotLocations}},
  \code{\link{addPlotObservations}},
  \code{\link{addSiteCharacteristics}},
  \code{\link{addSiteObservations}},
  \code{\link{addStratumObservations}},
  \code{\link{addSurfaceCoverObservations}},
  \code{\link{addTaxonBySiteData}}
}
