% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlotLocations.R
\name{addPlotLocations}
\alias{addPlotLocations}
\title{Adds/replaces plot location information}
\usage{
addPlotLocations(target, points, x, projectTitle, mapping,
  reset.locations = FALSE, missing.values = c(NA, ""), verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{points}{An object of class \code{\link{SpatialPoints}} with coordinates in a reference system.}

\item{x}{A data frame where each row corresponds to one plot observation. Columns can be varied.}

\item{projectTitle}{A string to identify the project title, which can be the same as one of the currently defined in \code{target}.}

\item{mapping}{A list with at least element name 'plotName', is used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Site variables that can be mapped are: 'authorLocation','locationNarrative', 'placeName', 'placeType'.
Additional optional mappings are: 'subPlotName'. Note that 'placeName' and 'placeType' will add new places to the list of locations}

\item{reset.locations}{Whether the 'locations' vector (a vector of place names) should be reset before adding new place names.}

\item{missing.values}{A character vector of values that should be considered as missing data.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds/replaces static plot location information (spatial coordinates, place names, ...) to plot elements of a VegX object from a data table where rows are plots,
using a mapping to identify plot and subplot (optional). Additional mapping elements are used to map specific variables.
}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
\code{\link{addSiteCharacteristics}}.
}
