% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlotLocations.R
\name{addPlotLocations}
\alias{addPlotLocations}
\title{Adds/replaces plot location information}
\usage{
addPlotLocations(target, x, mapping,
  proj4string = "+proj=longlat +ellps=WGS84", reset.places = FALSE,
  missing.values = c(NA, ""), verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one plot observation. Columns can be varied.}

\item{mapping}{A list with at least element name 'plotName', is used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Location variables that can be mapped are: 'x', 'y', 'authorLocation','locationNarrative', 'placeName', 'placeType'.
Additional optional mappings are: 'subPlotName'. Note that 'placeName' and 'placeType' will add new places to the list of places}

\item{proj4string}{A string with projection attributes (see \code{\link{proj4string}} of package \code{sp}) to be used when 'x' and 'y' are supplied. \
Note that coordinates will be transformed to "+proj=longlat +datum=WGS84".}

\item{reset.places}{Whether the 'places' vector should be reset before adding new place names.}

\item{missing.values}{A character vector of values that should be considered as missing data.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds/replaces static plot location information (spatial coordinates, place names, ...) to plot elements of a VegX object from a data table where rows are plots,
using a mapping to identify plot and subplot (optional). Additional mapping elements are used to map specific variables.
}
\examples{
data(mokihinui)

# Define location mapping
mapping = list(plotName = "Plot", subPlotName = "Subplot",
                  x = "Longitude", y = "Latitude")

# Create new Veg-X document with plot locations
x = addPlotLocations(newVegX(), moki_loc, mapping)

# Summary of the new Veg-X document
showElementTable(x, "plot")

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other add functions: \code{\link{addAggregateOrganismObservations}},
  \code{\link{addIndividualOrganismObservations}},
  \code{\link{addPlotObservations}},
  \code{\link{addSiteCharacteristics}},
  \code{\link{addSiteObservations}},
  \code{\link{addStratumObservations}},
  \code{\link{addSurfaceCoverObservations}},
  \code{\link{addTaxonBySiteData}}
}
