% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlotLocations.R
\name{addPlotLocations}
\alias{addPlotLocations}
\title{Adds/replaces plot location information}
\usage{
addPlotLocations(
  target,
  x,
  mapping,
  proj4string = NULL,
  reset.places = FALSE,
  toWGS84 = FALSE,
  methods = list(),
  missing.values = c(NA, ""),
  missing.coords = c(NA, 0, ""),
  missing.elevation = c(NA, 0, ""),
  verbose = TRUE
)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one plot. Columns can be varied.}

\item{mapping}{A named list whose elements are strings that correspond to column names in \code{x}. Names of the list should be:
\itemize{
   \item{\code{plotName} - A string identifying the vegetation plot within the data set (required).}
   \item{\code{subPlotName} - A string identifying a subplot of the plot given by \code{plotName} (optional).}
   \item{\code{x}, \code{y} - Spatial coordinates of the plot (optional).}
   \item{\code{elevation} - Elevation of the plot (optional).}
   \item{\code{authorLocation} - A string describing the location of the plot as made by the author (optional).}
   \item{\code{locationNarrative} -  (optional).}
   \item{\code{placeName}, \code{placeType} - A string of a place name and place type (e.g. province, county, ...) (optional).}
}
Note that \code{placeName} and \code{placeType} will add new places to the list of places.}

\item{proj4string}{A string with projection attributes (see \code{\link{proj4string}} of package \code{sp}) to be used when 'x' and 'y' are supplied. 
This parameter is needed if \code{toWGS84 = TRUE}.}

\item{reset.places}{Whether the 'places' vector should be reset before adding new place names.}

\item{toWGS84}{A boolean flag to indicate that coordinates should be transformed to "+proj=longlat +datum=WGS84".}

\item{methods}{A named list with measurement methods for plot horizontal/vertical location measurements (each being an object of class \code{\linkS4class{VegXMethodDefinition}}). 
Alternatively, methods can be specified using strings if predefined methods exist (see \code{\link{predefinedMeasurementMethod}}).
For example, \code{methods = c(xy = method1, elevation = method2)}. Measurement method for coordinates is not required, but that for 'elevation' is.}

\item{missing.values}{A character vector of values that should be considered as missing data (but see the following).}

\item{missing.coords}{A character vector of values that should be considered as missing coordinates (introduced to allow separate treatment).}

\item{missing.elevation}{A character vector of values that should be considered as missing elevation (introduced to allow separate treatment).}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds/replaces static plot location information (spatial coordinates, elevation, place names, ...) to plot elements of a VegX object from a data table where rows are plots.
}
\details{
Named elements in \code{mapping} other than those used by this function will be ignored. Missing value policy:
 \itemize{
    \item{Missing \code{plotName} values are interpreted as if the previous non-missing value has to be used to define plot.}
    \item{Missing \code{subPlotName} values are interpreted in that data refers to the parent plotName.}
    \item{Missing measurements (e.g. \code{elevation}, \code{x}, \code{y}, ...) are simply not added to the Veg-X document.}
 }
}
\examples{
data(mokihinui)

# Define location mapping
mapping = list(plotName = "Plot", x = "Longitude", y = "Latitude")

# Create new Veg-X document with plot locations
x = addPlotLocations(newVegX(), moki_loc, mapping,
                     proj4string = "+proj=longlat +datum=WGS84")

# Summary of the new Veg-X document
showElementTable(x, "plot")

# Add 'elevation' from another table (moki_site). This implies considering subplots.
mapping = list(plotName = "Plot", subPlotName = "Subplot", elevation = "Altitude")
x = addPlotLocations(x, moki_site, mapping, 
                     methods = list(elevation = "Elevation/m"))
                     
# Summary of the updated Veg-X document
showElementTable(x, "plot")

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other add functions: 
\code{\link{addAggregateOrganismObservations}()},
\code{\link{addCommunityObservations}()},
\code{\link{addIndividualOrganismObservations}()},
\code{\link{addPlotGeometries}()},
\code{\link{addPlotObservations}()},
\code{\link{addSiteCharacteristics}()},
\code{\link{addSiteObservations}()},
\code{\link{addStratumObservations}()},
\code{\link{addSurfaceCoverObservations}()},
\code{\link{addTaxonBySiteData}()}
}
\concept{add functions}
