% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlotObservations.R
\name{addPlotObservations}
\alias{addPlotObservations}
\title{Add plot observation records}
\usage{
addPlotObservations(
  target,
  x,
  mapping,
  missing.values = c(NA, ""),
  date.format = "\%Y-\%m-\%d",
  verbose = TRUE
)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one plot observation. Columns can be varied.}

\item{mapping}{A named list whose elements are strings that correspond to column names in \code{x}. Names of the list should be:
 \itemize{
   \item{\code{plotName} - A string identifying the vegetation plot within the data set (required).}
   \item{\code{subPlotName} - A string identifying a subplot of the plot given by \code{plotName} (optional).}
   \item{\code{obsStartDate} - Plot observation start date (required; see \code{date.format}).}
   \item{\code{projectTitle} - Title of the project related to the plot observation (optional).}
   \item{\code{obsEndDate} - Plot observation end date (optional; see  \code{date.format}).}
   \item{\code{observationParty} - Name of the party that undertook plot observation (optional).}
   \item{\code{plotUniqueIdentifier} - A string used to identify the plot uniquely, preferably globally unique (optional).}
}}

\item{missing.values}{A character vector of values that should be considered as missing data (see details).}

\item{date.format}{A character string specifying the input format of dates (see \code{\link{as.Date}}).}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds plot observation records to a VegX object from a data table where rows are plot observations.
}
\details{
Named elements in \code{mapping} beyond those used by this function will be ignored. Missing value policy:
 \itemize{
    \item{Missing \code{projectTitle}, \code{plotName}, \code{obsStartDate} or \code{obsEndDate} values are interpreted as if the previous non-missing value has to be used to define plot observation.}
    \item{Missing \code{subPlotName} values are interpreted in that data refers to the parent plotName.}
 }
}
\examples{
# Load source data
data(mokihinui)

# Define mapping
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")

# Create a new Veg-X document with projects, plots and plot observations (no data)
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)

# Examine the result
showElementTable(x, "plot")
showElementTable(x, "plotObservation")

# The same but capturing unique identifiers from data source IDs (not used by the Veg-X package)

mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate",
               plotUniqueIdentifier = "PlotID", plotObservationUniqueIdentifier = "PlotObsID")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)

showElementTable(x, "plot")
showElementTable(x, "plotObservation")
}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other add functions: 
\code{\link{addAggregateOrganismObservations}()},
\code{\link{addCommunityObservations}()},
\code{\link{addIndividualOrganismObservations}()},
\code{\link{addPlotGeometries}()},
\code{\link{addPlotLocations}()},
\code{\link{addSiteCharacteristics}()},
\code{\link{addSiteObservations}()},
\code{\link{addStratumObservations}()},
\code{\link{addSurfaceCoverObservations}()},
\code{\link{addTaxonBySiteData}()}
}
\concept{add functions}
