% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSiteCharacteristics.R
\name{addSiteCharacteristics}
\alias{addSiteCharacteristics}
\title{Adds/replaces site information}
\usage{
addSiteCharacteristics(
  target,
  x,
  mapping,
  measurementMethods = list(),
  missing.values = c(NA, ""),
  verbose = TRUE
)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one plot observation. Columns can be varied.}

\item{mapping}{A named list whose elements are strings that correspond to column names in \code{x}. Names of the list should be:
\itemize{
   \item{\code{plotName} - A string identifying the vegetation plot within the data set (required).}
   \item{\code{subPlotName} - A string identifying a subplot of the plot given by \code{plotName} (optional).}
   \item{\code{slope} - Slope of the plot (optional).}
   \item{\code{aspect} - Aspect (i.e. orientation) of the plot (optional).}
   \item{\code{landform} - Site land form (e.g. slope, ridge, saddle point; optional).}
   \item{\code{parentMaterial} - Underlying geological material (generally bedrock or a superficial or drift deposit) in which soil horizons form. (optional).}
}}

\item{measurementMethods}{A named list of objects of class \code{\linkS4class{VegXMethodDefinition}} with the measurement method
for each of the site variables stated in \code{mapping}. List names should be the same as the names of site variables
(e.g. \code{list(aspect = aspectMeth)} to specify the use of method \code{aspectMeth} for aspect measurements). Alternatively,
methods can be specified using strings if a predefined method exists (e.g. \code{list(aspect = "Aspect/degrees")}),
see \code{\link{predefinedMeasurementMethod}}.}

\item{missing.values}{A character vector of values that should be considered as missing observations/measurements.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds/replaces static site characteristics (topography, geology, ...) to plot elements of a VegX object from a data frame where rows are plots.
}
\details{
Named elements in \code{mapping} other than those used by this function will be ignored. Missing value policy:
 \itemize{
    \item{Missing \code{plotName} values are interpreted as if the previous non-missing value has to be used to define plot.}
    \item{Missing \code{subPlotName} values are interpreted in that data refers to the parent plotName.}
    \item{Missing measurements (e.g. \code{aspect}, \code{slope},...) are simply not added to the Veg-X document.}
 }
}
\examples{
# Load source data
data(mokihinui)

# Define mapping
sitemapping = list(plotName = "Plot", subPlotName = "Subplot",
                   slope = "PlotSlope", aspect = "PlotAspect")


# Create new Veg-X document with site characteristics
x = addSiteCharacteristics(newVegX(), moki_site, mapping = sitemapping,
                           measurementMethods = list(slope = "Slope/degrees",
                                                     aspect = "Aspect/degrees"))

# Inspect the result
showElementTable(x, "plot")

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other add functions: 
\code{\link{addAggregateOrganismObservations}()},
\code{\link{addCommunityObservations}()},
\code{\link{addIndividualOrganismObservations}()},
\code{\link{addPlotGeometries}()},
\code{\link{addPlotLocations}()},
\code{\link{addPlotObservations}()},
\code{\link{addSiteObservations}()},
\code{\link{addStratumObservations}()},
\code{\link{addSurfaceCoverObservations}()},
\code{\link{addTaxonBySiteData}()}
}
\concept{add functions}
