% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSiteCharacteristics.R
\name{addSiteCharacteristics}
\alias{addSiteCharacteristics}
\title{Adds/replaces site information}
\usage{
addSiteCharacteristics(target, x, projectTitle, mapping,
  measurementMethods = list(), missing.values = c(NA, ""), verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one plot observation. Columns can be varied.}

\item{projectTitle}{A string to identify the project title, which can be the same as one of the currently defined in \code{target}.}

\item{mapping}{A list with at least element name 'plotName', is used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Site variables that can be mapped are: 'elevation','slope', 'aspect', 'landform', 'parentMaterial', 'geologyClass'.
Additional optional mappings are: 'subPlotName'.}

\item{measurementMethods}{A named list of objects of class \code{\linkS4class{VegXMethod}} with the measurement method
for each of the abiotic variables stated in \code{mapping}. List names should be the same as abiotic variables
(e.g. \code{list(aspect = aspectMeth)} to specify the use of method '\code{aspectMeth}' for aspect measurements).}

\item{missing.values}{A character vector of values that should be considered as missing observations/measurements.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds/replaces static site characteristics (topography, geology, ...) to plot elements of a VegX object from a data table where rows are plots,
using a mapping to identify plot and subplot (optional). Additional mapping elements are used to map specific site variables.
}
\examples{
data(mokihinui)

# Create new Veg-X document
target = newVegX()

# Define mapping
sitemapping = list(plotName = "Plot", subPlotName = "Subplot",
                   elevation = "Altitude", slope = "PlotSlope", aspect = "PlotAspect")

# Define site methods
elevm = predefinedMeasurementMethod("Elevation meters")
slopeDeg = predefinedMeasurementMethod("Slope degrees")
aspectDeg = predefinedMeasurementMethod("Aspect degrees")

# Mapping process
w = addSiteCharacteristics(target, site, "Mokihinui",
                           mapping = sitemapping,
                           measurementMethods = list(elevation = elevm, slope = slopeDeg, aspect = aspectDeg))

summary(w)

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
\code{\link{addTaxonObservations}}, \code{\link{addTreeObservations}}.
}
