% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSiteObservations.R
\name{addSiteObservations}
\alias{addSiteObservations}
\title{Add site observation records}
\usage{
addSiteObservations(target, x, plotObservationMapping,
  soilMeasurementMapping = list(), climateMeasurementMapping = list(),
  waterBodyMeasurementMapping = list(),
  soilMeasurementMethods = list(), climateMeasurementMethods = list(),
  waterBodyMeasurementMethods = list(), date.format = "\%Y-\%m-\%d",
  missing.values = c(NA, ""), fill.methods = FALSE, verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one plot observation. Columns can be varied.}

\item{plotObservationMapping}{A list with element names 'plotName', 'obsStartDate', used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Additional optional mappings are: 'subPlotName'.}

\item{soilMeasurementMapping}{A named list used to specify the mapping of data columns to soil variables (e.g. soil1 = "pH" to map variable "pH" of the data frame into a first soil variable). List names should be the same as in \code{soilMeasurementMethods}.}

\item{climateMeasurementMapping}{A named list used to specify the mapping of data columns to climate variables. List names should be the same as in \code{climateMeasurementMethods}.}

\item{waterBodyMeasurementMapping}{A named list used to specify the mapping of data columns to water body variables. List names should be the same as in \code{waterBodyMeasurementMethods}.}

\item{soilMeasurementMethods}{A named list of objects of class \code{\linkS4class{VegXMethodDefinition}} with the measurement method
for each of the element names stated in \code{soilMeasurementMapping} (e.g. \code{list(soil1 = pHmeth)} to specify the use of method '\code{pHmeth}' for soil1). 
Alternatively, methods can be specified using strings if predefined methods exist (e.g. \code{list(soil1 = "pH/0-14")} to use the predefined method "pH/0-14"), 
see \code{\link{predefinedMeasurementMethod}}.}

\item{climateMeasurementMethods}{A named list of objects of class \code{\linkS4class{VegXMethodDefinition}} with the measurement method
for each of the element names stated in \code{climateMeasurementMapping}. List names should be the same as climate subject measurement variables. Alternatively,
methods can be specified using strings if predefined methods exist, 
see \code{\link{predefinedMeasurementMethod}}.}

\item{waterBodyMeasurementMethods}{A named list of objects of class \code{\linkS4class{VegXMethodDefinition}} with the measurement method
for each of the element names stated in \code{waterBodyMeasurementMapping}. List names should be the same as water body subject measurement variables. Alternatively,
methods can be specified using strings if predefined methods exist, 
see \code{\link{predefinedMeasurementMethod}}.}

\item{date.format}{A character string specifying the input format of dates (see \code{\link{as.Date}}).}

\item{missing.values}{A character vector of values that should be considered as missing observations/measurements.}

\item{fill.methods}{A flag to indicate that missing methods should be filled with dummy ones. This allows storing any environmental data, but without appropriate metadata.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds site observation records to a VegX object from a data table where rows are plot observations.
}
\details{
When defining methods, the users should preferably name subjects using the same strings as in predefined methods,
because this facilitates merging datasets where the same entities have been measured. 
The subjects currently recognized in predefined methods for soil observations are:
\itemize{
 \item{\code{pH}: Soil pH.}
 \item{\code{texture}: Soil texture class.}
}
For climate observations are:
\itemize{
 \item{\code{climate MAT}: Mean annual temperature.}
 \item{\code{climate TWM}: Temperature of the warmest month.}
 \item{\code{climate TCM}: Temperature of the coldest month.}
 \item{\code{climate MDR}: Mean diurnal range.}
 \item{\code{climate MAP}: Mean annual precipitation.}
 \item{\code{climate MAP}: Mean annual precipitation.}
 \item{\code{climate PDM}: Precipitation of the driest month.}
 \item{\code{climate PWM}: Precipitation of the wettest month.}
}
And for water body observations are:
\itemize{
 \item{\code{salinity}: Salinity of the water body.}
}
Please contact Veg-X developers to ask for additional subjects if you think they are rellevant for exchanging vegetation plot data. 

Missing value policy:
\itemize{
  \item{Missing 'plotName' or 'obsStartDate' values are interpreted as if the previous non-missing value has to be used to define plot observation.}
  \item{Missing 'subPlotName' values are interpreted in that observation refers to the parent plotName.}
  \item{Missing measurements are simply not added to the Veg-X document.}
}
}
\examples{
# Load source data
data(mokihinui)

# Define mapping
mapping = list(plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate")
# the element name refers to the subject and the string to the variable name              
soilmapping = list(soil1 = "pH") 

# Create new Veg-X document with site observations
# Uses predefined measurement method "pH/0-14"
x = addSiteObservations(newVegX(), moki_site,
                        plotObservationMapping = mapping,
                        soilMeasurementMapping = soilmapping,
                        soilMeasurementMethods = list(soil1 = "pH/0-14"))
# Examine results
summary(x)
head(showElementTable(x, "siteObservation", subject=TRUE))

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other add functions: \code{\link{addAggregateOrganismObservations}},
  \code{\link{addCommunityObservations}},
  \code{\link{addIndividualOrganismObservations}},
  \code{\link{addPlotGeometries}},
  \code{\link{addPlotLocations}},
  \code{\link{addPlotObservations}},
  \code{\link{addSiteCharacteristics}},
  \code{\link{addStratumObservations}},
  \code{\link{addSurfaceCoverObservations}},
  \code{\link{addTaxonBySiteData}}
}
\concept{add functions}
