% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSiteObservations.R
\name{addSiteObservations}
\alias{addSiteObservations}
\title{Add site observation records}
\usage{
addSiteObservations(target, x, plotObservationMapping,
  soilMeasurementMapping = list(), climateMeasurementMapping = list(),
  waterMassMeasurementMapping = list(), soilMeasurementMethods = list(),
  climateMeasurementMethods = list(), waterMassMeasurementMethods = list(),
  missing.values = c(NA, ""), verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one plot observation. Columns can be varied.}

\item{plotObservationMapping}{A list with element names 'plotName', 'obsStartDate', used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Additional optional mappings are: 'subPlotName'.}

\item{soilMeasurementMapping}{A list with element names equal to soil measurement subjects, used to specify the mapping of data columns (specified using strings for column names) onto these variables.}

\item{climateMeasurementMapping}{A list with element names equal to climate measurement subjects, used to specify the mapping of data columns (specified using strings for column names) onto these variables.}

\item{waterMassMeasurementMapping}{A list with element names equal to water mass measurement subjects, used to specify the mapping of data columns (specified using strings for column names) onto these variables.}

\item{soilMeasurementMethods}{A named list of objects of class \code{\linkS4class{VegXMethod}} with the measurement method
for each of the soil variables stated in \code{soilMeasurementMapping}. List names should be the same as soil subject measurement variables
(e.g. \code{list(pH = pHmeth)} to specify the use of method '\code{pHmeth}' for pH measurements).}

\item{climateMeasurementMethods}{A named list of objects of class \code{\linkS4class{VegXMethod}} with the measurement method
for each of the soil variables stated in \code{soilMeasurementMapping}. List names should be the same as climate subject measurement variables.}

\item{waterMassMeasurementMethods}{A named list of objects of class \code{\linkS4class{VegXMethod}} with the measurement method
for each of the soil variables stated in \code{soilMeasurementMapping}. List names should be the same as water mass subject measurement variables.}

\item{missing.values}{A character vector of values that should be considered as missing observations/measurements.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds site observation records to a VegX object from a data table where rows are plot observations,
using a mapping to identify plot observation: plot, subplot (optional) and observation date.
Additional mappings are used to map specific site variables.
}
\details{
Missing plotName or obsStartDate values are interpreted as if the previous non-missing value has to be used to define plot observation.
Missing subPlotName values are interpreted in that observation refers to the parent plotName.
Missing measurements are simply not added to the Veg-X document.
}
\examples{
# Load source data
data(mokihinui)

# Define mapping
mapping = list(plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate")
soilmapping = list(pH = "pH")

# Define pH method
pHMeth = predefinedMeasurementMethod("pH")

# Create new Veg-X document with site observations
x = addSiteObservations(newVegX(), moki_site,
                        plotObservationMapping = mapping,
                        soilMeasurementMapping = soilmapping,
                        soilMeasurementMethods = list(pH = pHMeth))
# Examine results
summary(x)
head(showElementTable(x, "siteObservation"))

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other add functions: \code{\link{addAggregateOrganismObservations}},
  \code{\link{addIndividualOrganismObservations}},
  \code{\link{addPlotLocations}},
  \code{\link{addPlotObservations}},
  \code{\link{addSiteCharacteristics}},
  \code{\link{addStratumObservations}},
  \code{\link{addTaxonBySiteData}}
}
