% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStratumObservations.R
\name{addStratumObservations}
\alias{addStratumObservations}
\title{Add stratum observation records}
\usage{
addStratumObservations(target, x, projectTitle, mapping, methods,
  stratumDefinition, missing.values = c(NA, ""), verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one aggregate taxon observation. Columns can be varied.}

\item{projectTitle}{A character string to identify the project title, which can be the same as one of the currently defined in \code{target}.}

\item{mapping}{A list with element names 'plotName', 'obsStartDate', and 'stratumName' used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Additional optional mappings are: 'obsEndDate', 'subPlotName', 'lowerLimitMeasurement', 'lowerLimitMeasurement', and mappings to other stratum measurements.}

\item{methods}{A list measurement methods for stratum measurements (an object of class \code{\linkS4class{VegXMethod}}).}

\item{stratumDefinition}{An object of class \code{\linkS4class{VegXStrata}} indicating the definition of strata.}

\item{missing.values}{A character vector of values that should be considered as missing observations/measurements.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds stratum observation records to a VegX object from a data table
}
\examples{
data(mokihinui)

# Create new Veg-X document
target = newVegX()

# Define mapping
mapping = list(plotName = "Plot", obsStartDate = "obsDate", stratumName = "Tier",
               lowerLimitMeasurement = "TierLower", upperLimitMeasurement = "TierUpper",
               cover = "CoverClass")


# Define cover scale
coverscale = definePlantCoverScale(name = "Recce cover scale", description = "Recce recording method by Allen",
                         citation = "Hurst, JM and Allen, RB. (2007) The Recce method for describing New Zealand vegetation – Field protocols. Landcare Research, Lincoln.",
                         breaks = c(0, 0.1, 1, 5, 25, 50, 75, 100),
                         midPoints = c(0.01, 0.05, 0.5, 15, 37.5, 62.5, 87.5),
                         values = c("P","1","2","3", "4", "5", "6"))

# Define height measurement methods
heightMethod = predefinedMeasurementMethod("Stratum height")

# Define strata
strataDef = defineCategoricalStrata(name = "Category strata",
                                 description = "Strata with boundaries undefined",
                                 strataNames = paste0("Tier ",1:7))

# Mapping process
x = addStratumObservations(target, tier, "Mokihinui",
                        mapping = mapping,
                        methods = list(lowerLimitMeasurement = heightMethod,
                                       upperLimitMeasurement = heightMethod,
                                       cover=coverscale),
                        stratumDefinition = strataDef)

summary(x)

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other add functions: \code{\link{addAggregateOrganismObservations}},
  \code{\link{addIndividualOrganismObservations}},
  \code{\link{addPlotLocations}},
  \code{\link{addSiteCharacteristics}},
  \code{\link{addSiteObservations}},
  \code{\link{addTaxonBySiteData}}
}
