% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSurfaceCoverObservations.R
\name{addSurfaceCoverObservations}
\alias{addSurfaceCoverObservations}
\title{Add surface cover observation records}
\usage{
addSurfaceCoverObservations(target, x, mapping, coverMethod,
  surfaceTypeDefinition, missing.values = c(NA, ""), verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one surface cover observation (e.g. bare rock percent cover). Columns can be varied.}

\item{mapping}{A list with element names 'plotName',  'obsStartDate', 'surfaceName' and 'coverMeasurement'
are used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Additional optional mappings are: 'subPlotName'.}

\item{coverMethod}{A method definition for surface cover measurements (an object of class \code{\linkS4class{VegXMethodDefinition}}).}

\item{surfaceTypeDefinition}{An object of class \code{\linkS4class{VegXSurfaceTypeDefinition}} indicating the definition of surface types.}

\item{missing.values}{A character vector of values that should be considered as missing observations/measurements.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds surface cover observation records to a VegX object from a data table where rows are cover observations.
}
\details{
Missing value policy:
 \itemize{
   \item{Missing 'plotName', 'obsStartDate' or 'surfaceName' values are interpreted as if the previous non-missing value has to be used to define plot observation.}
   \item{Missing 'subPlotName' values are interpreted in that observation refers to the parent plotName.}
   \item{Missing measurements are simply not added to the Veg-X document.}
}
}
\examples{
# Load source data
data(mtfyffe)

# Examine surface cover data
head(mtfyffe_groundcover)

# Define mapping
mapping = list(plotName = "Plot", obsStartDate = "PlotObsStartDate",
               surfaceName = "PlotGroundCover", coverMeasurement = "Value")


# Get cover measurement method from predefined list
coverMethod = predefinedMeasurementMethod("Surface cover/\%")

# Define surface types from the data
unique(mtfyffe_groundcover$PlotGroundCover)
surfaceTypes = defineSurfaceTypes(name = "Default surface types",
                                  description = "Five surface categories",
                                  surfaceNames = c("Vegetation", "Moss", "Litter", "Exposed Soil", "Rock"))

# Create new Veg-X document with surface cover observations
x = addSurfaceCoverObservations(newVegX(), mtfyffe_groundcover, mapping,
                                coverMethod, surfaceTypes)

# Examine results
head(showElementTable(x, "surfaceCoverObservation"))


# Another example with different surface types
data(takitimu)
head(taki_groundcover)

unique(taki_groundcover$PlotGroundCover)
surfaceTypes = defineSurfaceTypes(name = "Default surface types",
                              description = "Five surface categories",
                              surfaceNames = c("Vegetation", "Soil", "Erosion Pavement", "Litter","Rock"))

x = addSurfaceCoverObservations(newVegX(), taki_groundcover, mapping,
                                coverMethod, surfaceTypes)

head(showElementTable(x, "surfaceCoverObservation"))

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other add functions: \code{\link{addAggregateOrganismObservations}},
  \code{\link{addIndividualOrganismObservations}},
  \code{\link{addPlotGeometries}},
  \code{\link{addPlotLocations}},
  \code{\link{addPlotObservations}},
  \code{\link{addSiteCharacteristics}},
  \code{\link{addSiteObservations}},
  \code{\link{addStratumObservations}},
  \code{\link{addTaxonBySiteData}}
}
