% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTaxonBySiteData.R
\name{addTaxonBySiteData}
\alias{addTaxonBySiteData}
\title{Add a Taxon by Site table}
\usage{
addTaxonBySiteData(target, x, projectTitle,
  method = defaultPercentCoverMethod(), obsDates = Sys.Date(),
  absence.values = c(NA, 0), verbose = TRUE)
}
\arguments{
\item{target}{the original object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{site-by-species releve table}

\item{projectTitle}{a character string to identify the project title, which can be the same as one of the currently defined in \code{target}.}

\item{method}{measurement method for aggregated plant abundance (an object of class \code{\linkS4class{VegXMethod}}).}

\item{obsDates}{a vector of \code{\link{Date}} objects with plot observation dates.}

\item{absence.values}{a vector of values to be interpreted as missing plant information.}

\item{verbose}{flag to indicate console output of the data integration process.}
}
\value{
an object of class \code{\linkS4class{VegX}}
}
\description{
Adds the aggregated organism observations of a taxon-by-site table to an existing VegX object.
Data can be added to an existing project or a new project can be defined within the VegX object,
depending on the input project title.
Vegetation plots and taxon names can be the same as those already existing in the target VegX.
Plot observations, however, are always considered new.
}
