% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTaxonObservationRecords.R
\name{addTaxonObservationRecords}
\alias{addTaxonObservationRecords}
\title{Add a taxon observation records}
\usage{
addTaxonObservationRecords(target, x, projectTitle, mapping,
  abundanceMethod = defaultPercentCoverMethod(),
  stratumDefinition = stratumDefinitionByHeight(), verbose = TRUE)
}
\arguments{
\item{target}{the original object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{a data frame where each row corresponds to one aggregated taxon observation. Columns can be varied.}

\item{projectTitle}{A character string to identify the project title, which can be the same as one of the currently defined in \code{target}.}

\item{mapping}{A list with element names 'plotName', 'obsStartDate', 'taxonAuthorName' and 'value', used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Additional optional mappings are: 'obsEndDate' and 'stratumName'.}

\item{abundanceMethod}{Measurement method for aggregated plant abundance (an object of class \code{\linkS4class{VegXMethod}}).}

\item{stratumDefinition}{an object of class \code{\linkS4class{VegXStrata}} indicating the definition of strata.}

\item{verbose}{flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}
}
\description{
Adds aggregated taxon observation records to a VegX object from a data table
using a mapping to identify columns: plot, observation date, stratum, taxon name and value.
}
