% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTaxonObservationRecords.R
\name{addTaxonObservationRecords}
\alias{addTaxonObservationRecords}
\title{Add a taxon observation records}
\usage{
addTaxonObservationRecords(target, x, projectTitle, obsMapping,
  abundanceMethod = defaultPercentCoverMethod(), strMapping = NULL,
  stratumDefinition = defaultStratumDefinition())
}
\arguments{
\item{target}{the original object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{a data frame where each row corresponds to one aggregated taxon observation. Columns can be varied.}

\item{projectTitle}{a character string to identify the project title, which can be the same as one of the currently defined in \code{target}.}

\item{obsMapping}{a character vector with element names 'plotName', 'obsStartDate', 'stratumName', 'taxonAuthorName' and 'value'.}

\item{abundanceMethod}{measurement method for aggregated plant abundance (an object of class \code{\linkS4class{VegXMethod}}).}

\item{stratumMethod}{}
}
\value{
The modified object of class \code{\linkS4class{VegX}}
}
\description{
Adds aggregated taxon observation records to a VegX object from a data table
using a mapping to identify columns: plot, observation date, stratum, taxon name and value.
}
