% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTreeObservationRecords.R
\name{addTreeObservationRecords}
\alias{addTreeObservationRecords}
\title{Add tree observation records}
\usage{
addTreeObservationRecords(target, x, projectTitle, mapping, diameterMethod,
  heightMethod = NULL, stratumDefinition = NULL, missing.values = c(NA,
  "0", ""), verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{x}{A data frame where each row corresponds to one tree observation. Columns can be varied.}

\item{projectTitle}{A string to identify the project title, which can be the same as one of the currently defined in \code{target}.}

\item{mapping}{A list with element names 'plotName', 'obsStartDate', 'taxonAuthorName' and 'diameter', used to specify the mapping of data columns (specified using strings for column names) onto these variables.
Additional optional mappings are: 'subPlotName', 'obsEndDate', 'individual', 'height' and 'stratumName'.}

\item{stratumDefinition}{An object of class \code{\linkS4class{VegXStrata}} indicating the definition of strata.}

\item{missing.values}{A character vector of values that should be considered as missing observations/measurements.}

\item{verbose}{A boolean flag to indicate console output of the data integration process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Adds tree observation records to a VegX object from a data table,
using a mapping to identify columns: plot, observation date, taxon name and diameter.
Additional mappings can be used to specify a stratum where the tree is located.
}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
\code{\link{addTaxonObservationRecords}}.
}
