% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildVegX.R
\name{buildVegX}
\alias{buildVegX}
\title{Convert Data to a VegX Object}
\usage{
buildVegX(
  input.info = NULL,
  fill.project = TRUE,
  add.people = TRUE,
  add.citation = TRUE,
  proj.string = "+proj=longlat +datum=WGS84"
)
}
\arguments{
\item{input.info}{a list. The output of the VegX high-level function
function `prepareInputData()`.}

\item{fill.project}{logical. Should the information about the project be
added to the VegX object? Defaults to TRUE.}

\item{add.people}{logical. Should the information about the people involved
in the project be added to the VegX object? Defaults to TRUE.}

\item{add.citation}{logical. Should the information about the literature
associated with the vegetation data be added to the VegX object? Defaults
to TRUE.}

\item{proj.string}{a string with the projection attributes of the plot/site
geographical coordinates to be passed to the function `addPlotLocations()`
(see `proj4string` of package `sp`).}
}
\value{
The ...

A list with ...
}
\description{
This function ...
}
\details{


This function ...
}
\seealso{
Other high-level functions: 
\code{\link{prepareInputData}()},
\code{\link{prepareMapping}()}
}
\concept{high-level functions}
