% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineStrata.R
\name{defineCategoricalStrata}
\alias{defineCategoricalStrata}
\title{Strata definition by simple category}
\usage{
defineCategoricalStrata(name = "Strata by categories",
  description = "Vegetation categorical strata", citationString = "",
  DOI = "", strataNames = c("s1", "s2", "s3"), strataDefinitions = NULL)
}
\arguments{
\item{name}{A string to identify the stratum definition.}

\item{description}{A string describing how strata are defined.}

\item{citationString}{A string with the bibliographic reference for the method.}

\item{DOI}{A string with the DOI of a resource describing the method.}

\item{strataNames}{A numeric vector of stratum codes (of length equal to the number of strata).}
}
\value{
An object of class \code{\linkS4class{VegXStrataDefinition}}
}
\description{
Strata definition by simple category
}
\seealso{
Other define strata functions: \code{\link{defineHeightStrata}},
  \code{\link{defineMixedStrata}}
}
