% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineStrata.R
\name{defineHeightStrata}
\alias{defineHeightStrata}
\title{Stratum definition by height}
\usage{
defineHeightStrata(
  name = "Strata by height",
  description = "Vegetation strata defined by height in m",
  citationString = "",
  DOI = "",
  heightBreaks = c(0, 1, 3, 5),
  heightUnit = "m",
  strataNames = c("s1", "s2", "s3"),
  strataDefinitions = NULL
)
}
\arguments{
\item{name}{A string to identify the stratum definition.}

\item{description}{A string describing how strata are defined.}

\item{citationString}{A string with the bibliographic reference for the
method.}

\item{DOI}{A string with the DOI of a resource describing the method.}

\item{heightBreaks}{A numeric vector with height limits between strata (of
length equal to the number of strata plus one).}

\item{heightUnit}{A string to identify height units.}

\item{strataNames}{A character vector of stratum codes (of length equal to
the number of strata).}

\item{strataDefinitions}{A character vector with strata definitions}
}
\value{
An object of class \code{\linkS4class{VegXStrataDefinition}}
}
\description{
Height-based strata definition
}
\seealso{
Other define strata functions: 
\code{\link{defineCategoricalStrata}()},
\code{\link{defineMixedStrata}()}
}
\concept{define strata functions}
