% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineStrata.R
\name{defineMixedStrata}
\alias{defineMixedStrata}
\title{Mixed stratum definition}
\usage{
defineMixedStrata(
  name = "Strata by height or category",
  description = paste0("Vegetation strata defined by height in m",
    " and other strata by category"),
  citationString = "",
  DOI = "",
  heightStrataBreaks = c(0, 1, 3, 5),
  heightStrataUnit = "m",
  heightStrataNames = c("s1", "s2", "s3"),
  categoryStrataNames = "s4",
  heightStrataDefinitions = NULL,
  categoryStrataDefinitions = NULL,
  order = NULL
)
}
\arguments{
\item{name}{A string to identify the stratum definition.}

\item{description}{A string describing how strata are defined.}

\item{citationString}{A string with the bibliographic reference for the
method.}

\item{DOI}{A string with the DOI of a resource describing the method.}

\item{heightStrataBreaks}{A numeric vector with height limits between height
strata (of length equal to the number of height strata plus one).}

\item{heightStrataUnit}{A string to identify height units.}

\item{heightStrataNames}{A numeric vector of stratum codes (of length equal
to the number of height strata).}

\item{categoryStrataNames}{A numeric vector of categorical stratum codes (of
length equal to the number of categorical strata).}

\item{heightStrataDefinitions}{A character vector with height strata
definitions}

\item{categoryStrataDefinitions}{A character vector with category strata
definitions}

\item{order}{A numeric vector to specify order strata (indices starting from
height strata and continuing with category strata).}
}
\value{
An object of class \code{\linkS4class{VegXStrataDefinition}}
}
\description{
Define some strata by height and other by category
}
\examples{
strataDef = defineMixedStrata(name = "Recce strata",
                 description = "Standard Recce stratum definition",
                 citation = "Hurst, JM and Allen, RB. (2007) 
                     The Recce method for describing New Zealand vegetation – Field protocols. 
                     Landcare Research, Lincoln.",
                 heightStrataBreaks = c(0, 0.3,2.0,5, 12, 25, 50),
                 heightStrataNames = paste0("Tier ",1:6),
                 categoryStrataNames = "Tier 7")

}
\seealso{
Other define strata functions: 
\code{\link{defineCategoricalStrata}()},
\code{\link{defineHeightStrata}()}
}
\concept{define strata functions}
