% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineQualitativeScaleMethod.R
\name{defineQualitativeScaleMethod}
\alias{defineQualitativeScaleMethod}
\title{Qualitative scale method definition}
\usage{
defineQualitativeScaleMethod(name, description, subject, codes, citation = "",
  definitions = NULL)
}
\arguments{
\item{name}{String with the name of the method.}

\item{description}{String describing the method.}

\item{subject}{A string to identify the subject.}

\item{codes}{A character vector of class codes.}

\item{citation}{A string with the bibliographic reference for the method.}

\item{definitions}{A character vector of class definitions.}
}
\value{
an object of class \code{\linkS4class{VegXMethod}}
}
\description{
A general function to define a method for measurements of a given subject in a qualitative (i.e. categorical) scale.
}
\examples{

#Qualitative scale with three levels
defineQualitativeScaleMethod("scale1", "Description for scale1", "subject1",
                             codes = c("A", "B", "C"))




}
\seealso{
Other define measurement functions: \code{\link{defineOrdinalScaleMethod}},
  \code{\link{defineQuantitativeScaleMethod}},
  \code{\link{predefinedMeasurementMethod}}
}
