% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineSurfaceCategories.R
\name{defineSurfaceCategories}
\alias{defineSurfaceCategories}
\title{Surface type definition by simple category}
\usage{
defineSurfaceCategories(name, description, surfaceNames, definitions = NULL,
  citation = "")
}
\arguments{
\item{name}{A string to identify the surface type definition.}

\item{description}{A string describing how surface types are defined.}

\item{surfaceNames}{A character vector of names for surface types.}

\item{definitions}{A character vector of definition of surface types.}

\item{citation}{A string with the bibliographic reference for the method.}
}
\value{
an object of class \code{\linkS4class{VegXSurfaceTypeDefinition}}
}
\description{
Surface type definition by simple category
}
\examples{

defineSurfaceCategories(name = "Surface classes",
                        description = "Four simple surface categories",
                        surfaceNames = c("bare soil", "water", "rocks", "vegetation"))

}
