% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineStrataByHeight.R
\name{defineSurfaceCategories}
\alias{defineSurfaceCategories}
\title{Surface cover definition by simple category}
\usage{
defineSurfaceCategories(name = "Surface covers",
  description = "Four simple surface categories",
  surfaceNames = c("bare soil", "water", "rock", "vegetation"))
}
\arguments{
\item{name}{A string to identify the surface cover definition.}

\item{description}{A string describing how surface covers are defined.}

\item{surfaceNames}{A numeric vector of surface codes (of length equal to the number of strata).}
}
\value{
an object of class \code{\linkS4class{VegXStrata}}
}
\description{
Surface cover definition by simple category
}
