% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeVegX.R
\name{mergeVegX}
\alias{mergeVegX}
\title{Merge Veg-X documents}
\usage{
mergeVegX(x, y, verbose = TRUE)
}
\arguments{
\item{x, y}{The objects of class \code{\linkS4class{VegX}} to be merged}

\item{verbose}{A flag to indicate console output of the data integration process.}
}
\value{
An object of class \code{\linkS4class{VegX}} with the pooled data
}
\description{
Merges two Veg-X documents while considering that some of their entities may be shared.
}
\details{
Some entities are attempted to be merged or are kept as separate entities depeding on their definition:
\itemize{
  \item \code{projects} are merged when their element \code{title} has the same value
  \item \code{plots} are merged when their element \code{plotName} has the same value
  \item \code{plotObservations} are merged when both their \code{plotID} and \code{obsStartDate} elements have the same value
  \item \code{taxonNameUsageConcepts} are merged when their element \code{authorTaxonName} has the same value
  \item \code{methods} are merged when their element \code{name} has the same value
  \item \code{strata} are merged when their element \code{name} has the same value
  \item \code{stratumObservations} are merged when both their \code{stratumID} and \code{plotObservationID} elements have the same value
  \item \code{aggregateOrganismObservations} are merged when their \code{plotObservationID} and \code{taxonUsageConceptID} (and \code{stratumObservationID}, if defined) have the same value
  \item \code{individualOrganisms} are merged when both their \code{plotID} and \code{identificationLabel} have the same value
  \item \code{individualOrganismObservations} are merged when both their \code{plotObservationID} and \code{individualOrganismID} have the same value.
  \item \code{siteObservations} are merged into the same element when their element \code{plotObservationID} has the same value, but particular measurements are always added
  as if they were distinct pieces of information.
}
Merging to these entities may cause interruption of the process if the two entities to be merged
have different value for the same element. Other entities (e.g., \code{attributes} of a method) are always considered as distinct
entities between the two data sets to be merged and hence are simply copied to the result.
}
\examples{

data(mokihinui)

# Create document 'x' with aggregate taxon observations
taxmapping = list(plotName = "Plot", obsStartDate = "obsDate", authorTaxonName = "PreferredSpeciesName",
              stratumName = "Tier", cover = "Category")
coverscale = definePlantCoverScale(name = "Recce cover scale", description = "Recce recording method by Allen",
                         citation = "Hurst, JM and Allen, RB. (2007) The Recce method for describing New Zealand vegetation – Field protocols. Landcare Research, Lincoln.",
                         breaks = c(0, 0.1, 1, 5, 25, 50, 75, 100),
                         midPoints = c(0.01, 0.05, 0.5, 15, 37.5, 62.5, 87.5),
                         values = c("P","1","2","3", "4", "5", "6"))
strataDef = defineMixedStrata(name = "Recce strata",
                              description = "Standard Recce stratum definition",
                              citation = "Hurst, JM and Allen, RB. (2007) The Recce method for describing New Zealand vegetation – Field protocols. Landcare Research, Lincoln.",
                              heightStrataBreaks = c(0, 0.3,2.0,5, 12, 25, 50),
                              heightStrataNames = paste0("Tier ",1:6),
                              categoryStrataNames = "Tier 7",
                              categoryStrataDefinition = "Epiphytes")
x = addAggregateOrganismObservations(newVegX(), tcv, "Mokihinui",
                        mapping = taxmapping,
                        methods = c(cover=coverscale),
                        stratumDefinition = strataDef)

# Create document 'y' with tree observations
treemapping = list(plotName = "Plot", subPlotName = "Subplot", obsStartDate = "obsDate",
                   authorTaxonName = "PreferredSpeciesName", diameterMeasurement = "Diameter")
diamMeth = predefinedMeasurementMethod("DBH")
y = addIndividualOrganismObservations(newVegX(), dia, "Mokihinui",
                        mapping = treemapping,
                        methods = c(diameterMeasurement = diamMeth))

# Merge 'x' and 'y'
z = mergeVegX(x,y)
summary(z)

}
\seealso{
\code{\linkS4class{VegX}}
}
