% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeVegX.R
\name{mergeVegX}
\alias{mergeVegX}
\title{Merges two Veg-X documents}
\usage{
mergeVegX(x, y, verbose = TRUE)
}
\arguments{
\item{x, y}{The objects of class \code{\linkS4class{VegX}} to be merged}

\item{verbose}{A flag to indicate console output of the data integration process.}
}
\value{
An object of class \code{\linkS4class{VegX}} with the pooled data
}
\description{
Merges two Veg-X documents while considering that some entities may be shared.
}
\details{
Some entities are attempted to be merged or are kept as separate entities depeding on their definition:
\itemize{
  \item \code{projects} are merged when their element \code{title} has the same value
  \item \code{plots} are merged when their element \code{plotName} has the same value
  \item \code{plotObservations} are merged when both \code{plotID} and \code{obsStartDate} have the same value
  \item \code{taxonNameUsageConcepts} are merged when element \code{authorName} has the same value
  \item \code{methods} are merged when element \code{name} has the same value
}
Merging to these entities may cause interruption of the process if the two entities to be merged
have different value for the same element. Other entities (\code{stratumObservation}, \code{individualOrganismObservation}) are always considered as distinct
entities between the two data sets to be merged and hence are simply copied to the result.
}
