% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationDatasets.r
\docType{data}
\name{mokihinui}
\alias{mokihinui}
\title{Mokihinui example dataset for Veg-X}
\description{
The data set concerns observations made on 5 plots with 4 subplots each.
The data set includes four data frames:
}
\details{
\itemize{
 \item moki_loc Location of vegetation plots (5 records x 25 variables)
 \item moki_site Site characteristics of plots/subplots (25 records x 43 variables)
 \item moki_tcv Observations of cover of taxon categories, grouped by vegetation strata (578 records x 33 variables)
 \item moki_dia Observations made on individual trees (968 records x 33 variables)
}
}
\references{
Wiser, S.K., P.J. Bellingham and L.E. Burrows.  2001.  Managing biodiversity information:  development of the National Vegetation Survey Databank.  New Zealand Journal of Ecology 25(2): 1-17.
}
\author{
NVS database. Landcare Research (New Zealand) \email{nvs@landcareresearch.co.nz}
}
\keyword{data}
