% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predefinedMeasurementMethod.R
\name{predefinedMeasurementMethod}
\alias{predefinedMeasurementMethod}
\title{Predefined measurement methods}
\usage{
predefinedMeasurementMethod(name)
}
\arguments{
\item{name}{A string with the desired measurement method or a subject (in this case, the default method is returned). 
The list of available predefined methods and subjects is found in \code{\link{quantitative_methods}}.}
}
\value{
an object of class \code{\linkS4class{VegXMethodDefinition}}
}
\description{
Creates a measurement method for Veg-X documents from a set of predefined options
}
\examples{
# Create a method for plant percent cover
predefinedMeasurementMethod("Plant cover/\%")

}
\seealso{
Other define measurement functions: 
\code{\link{defineOrdinalScaleMethod}()},
\code{\link{defineQualitativeScaleMethod}()},
\code{\link{defineQuantitativeScaleMethod}()}
}
\concept{define measurement functions}
