% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predefinedMeasurementMethod.R
\name{predefinedMeasurementMethod}
\alias{predefinedMeasurementMethod}
\title{Predefined measurement methods}
\usage{
predefinedMeasurementMethod(name)
}
\arguments{
\item{name}{A string with the desired measurement method. Current accepted options are:
\itemize{
 \item{\code{"Plant cover/\%"}: Plant cover as percentage of ground covered by the projection.}
 \item{\code{"Plant counts"}: Number of individuals in the plot.}
 \item{\code{"Plant frequency/\%"}: Frequency of occurrence in fixed subunits of the plot.}
 \item{\code{"Stratum height/m"}: Stratum height in meters above the ground.}
 \item{\code{"Stratum height/cm"}: Stratum height in cm above the ground.}
 \item{\code{"Plant height/m"}: Plant height in meters above the ground.}
 \item{\code{"Plant height/cm"}: Plant height in cm above the ground.}
 \item{\code{"DBH/cm"}: Diameter at breast height, in cm.}
 \item{\code{"Surface cover/\%"}: Surface covered by a surface type as percentage of ground covered by the projection.}
 \item{\code{"pH"}: pH scale from 0 to 14.}
 \item{\code{"Slope/degrees"}: Slope measured using degrees.}
 \item{\code{"Aspect/degrees"}: Aspect measured using degrees from North.}
 \item{\code{"Elevation/m"}: Elevation measured using meters above sea level.}
}}
}
\value{
an object of class \code{\linkS4class{VegXMethod}}
}
\description{
Creates a measurement method for Veg-X documents from a set of predefined options
}
\examples{
# Create a method for plant percent cover
predefinedMeasurementMethod("Plant cover/\%")

}
\seealso{
Other define measurement functions: \code{\link{defineOrdinalScaleMethod}},
  \code{\link{defineQualitativeScaleMethod}},
  \code{\link{defineQuantitativeScaleMethod}}
}
