% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predefinedMeasurementMethod.R
\name{predefinedMeasurementMethod}
\alias{predefinedMeasurementMethod}
\title{Predefined measurement methods}
\usage{
predefinedMeasurementMethod(name)
}
\arguments{
\item{name}{A string with the desired measurement method. Current options are:
\code{"Percent cover"}, \code{"Individual counts"}, \code{"Frequency of occurrence"}}
}
\value{
an object of class \code{\linkS4class{VegXMethod}}
}
\description{
Predefined measurement methods
}
