% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predefinedMeasurementMethod.R
\name{predefinedMeasurementMethod}
\alias{predefinedMeasurementMethod}
\title{Predefined measurement methods}
\usage{
predefinedMeasurementMethod(name)
}
\arguments{
\item{name}{A string with the desired measurement method. Current accepted options are:
\code{"Percent cover"}, \code{"Individual counts"}, \code{"Frequency of occurrence"},
\code{"Plant height"}, \code{"DBH"}, \code{"pH"}, \code{"Slope degrees"}, \code{"Aspect degrees"},
\code{"Elevation meters"}.}
}
\value{
an object of class \code{\linkS4class{VegXMethod}}
}
\description{
Creates a measurement method for Veg-X documents from a set of predefined options
}
\examples{
# Create a method for plant percent cover
predefinedMeasurementMethod("Percent cover")

}
