% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareInputData.R
\name{prepareInputData}
\alias{prepareInputData}
\title{Prepare Data to the VegX Format}
\usage{
prepareInputData(
  plot.info = NULL,
  plot.obs = NULL,
  site.data = NULL,
  surface.data = NULL,
  community.data = NULL,
  species.data = NULL,
  individual.data = NULL,
  project.info = NULL,
  people.info = NULL,
  citation.info = NULL,
  mapping = NULL,
  project.title = NULL,
  miss.fields = TRUE,
  fixed.date = Sys.Date(),
  spp.names = c("organismNameMorpho", "organismNameOriginal", "organismName"),
  summarise.data = TRUE
)
}
\arguments{
\item{plot.info}{a data frame. General information of the plots or invariant
inventory site characteristics (e.g. plot or site codes, names and
identifiers, methods, location (place names), geographical coordinates,
elevation, aspect, slope). If the site characteristic can vary between
census, it should be provided as `site.data`.}

\item{plot.obs}{a data frame. Observations or census information related to
the plots or inventories (e.g. census dates, data generator, citation,
licenses, taxonomic quality checks)}

\item{site.data}{a data frame. Observations or measurements associated with
the inventory site that may vary in time (soil and climate conditions,
legal protection, land use, vegetation type)}

\item{surface.data}{a data frame. Surface cover measurements associated with
the plot at a given time period.}

\item{community.data}{a data frame. Observations or measurements associated
with the entire vegetation, i.e. observations at the stand or community
level (e.g. density, biomass, species diversity, canopy cover, vegetation
height, successional status)}

\item{species.data}{a data frame. The aggregated observations per species
(e.g. species name, count of individuals, total biomass, frequency)}

\item{individual.data}{a data frame. The observations per plant individual
(e.g. species name, field tag, diameter, height, coordinates within the
plot)}

\item{project.info}{a data frame. The information about the project that
generated/compiled the vegetation data (e.g. abstract, funding, citation
and short descriptions of the studied area and the typical project design)}

\item{people.info}{a data frame. The information about the people involved in
the project that generated/compiled the vegetation data (e.g. full names,
roles, addresses, email)}

\item{citation.info}{a data frame. The information about the literature
associated with the vegetation data (e.g. citation ID, authors, year,
title, etc.)}

\item{mapping}{a named list with no default. Usually the output of the VegX
function function `prepareMapping()`.}

\item{project.title}{a character with the project name/title. Only needed if
the user wants to use a different name than the one provided in the
argument `mapping`.}

\item{miss.fields}{logical. Should any missing fields be auto-completed in
the field mapping from matching the names of the columns in the input data?
Defaults to TRUE.}

\item{fixed.date}{date. A fixed date to be added for projects or plots
without census date information. Defaults to the current system data (i.e.
`Sys.Date()`). For dates other than the default, preferably use the
following format: "YYYY-MM-DD".}

\item{spp.names}{character vector. The priority order for defining the
working organism name for building the VegX object and for calculating
species diversity indices. Defaults to the following order:
'organismNameMorpho', 'organismNameOriginal' and 'organismName'.}

\item{summarise.data}{logical. Should individual or species plot data be
aggregated at higher levels of information (i.e. species and community
levels)? Defaults to TRUE.}
}
\value{
A list with ...
}
\description{
This function ...
}
\details{
This function
}
\seealso{
Other high-level functions: 
\code{\link{buildVegX}()},
\code{\link{prepareMapping}()}
}
\concept{high-level functions}
