% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevetable2VegX.R
\name{relevetable2VegX}
\alias{relevetable2VegX}
\title{Create a Veg-X object from a releve table}
\usage{
relevetable2VegX(x, method = defaultPercentCoverMethod(),
  obsDates = Sys.Date(), absence.values = c(NA, 0))
}
\arguments{
\item{x}{site-by-species releve table}

\item{method}{measurement method for aggregated plant abundance}

\item{obsDates}{a vector of \code{\link{Date}} objects with plot observation dates.}

\item{absence.values}{a vector of values to be interpreted as missing plant information}
}
\value{
an object of class \code{\linkS4class{VegX}}
}
\description{
Create a Veg-X object from a releve table
}
