% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOriginalIdentificationConcepts.R
\name{setOriginalIdentificationConcepts}
\alias{setOriginalIdentificationConcepts}
\title{Sets original identification taxon concepts}
\usage{
setOriginalIdentificationConcepts(
  target,
  citationStringAll = "",
  x = NULL,
  mapping = list(),
  date.format = "\%Y-\%m-\%d",
  missing.values = c(NA, ""),
  verbose = TRUE
)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified}

\item{citationStringAll}{A string with the bibliographic citation to be applied to all organism identities of the VegX object (using the original organism names as taxon names),
or to all original organism names listed in \code{x}.}

\item{x}{A data frame where each row corresponds to one organism identity, given by a column that can be mapped to \code{originalOrganismName}.}

\item{mapping}{A named list whose elements are strings that correspond to column names in \code{x}. Names of the list should be:
\itemize{
  \item{\code{originalOrganismName} - A string with the original name given by the author of the data set (required).}
  \item{\code{conceptName} - A string with the taxon name forming the taxon concept, if different from \code{originalOrganismName} (optional).}
  \item{\code{conceptCitation} - A string with the bibliographic citation forming the taxon concept (optional).}
  \item{\code{assertionDate} - Date of taxon concept assertion (see \code{date.format}) (optional).}
  \item{\code{assertionParty} - Name of the party that undertook taxon concept assertion (optional).}
}}

\item{date.format}{A character string specifying the input format of dates (see \code{\link{as.Date}}).}

\item{missing.values}{A character vector of values that should be considered as missing data (see details).}

\item{verbose}{A boolean flag to indicate console output of the concept identification process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Sets (or resets) taxon concepts associated with the original identification of organisms
}
\examples{

# Load source data
data(mokihinui)

# Create new Veg-X document with aggregate organism observations
mapping = list(plotName = "Plot", obsStartDate = "PlotObsStartDate",
               taxonName = "NVSSpeciesName",
               stratumName = "Tier", cover = "Category")
coverscale = defineOrdinalScaleMethod(name = "Recce cover scale",
                   description = "Recce recording method by Hurst/Allen",
                   subject = "plant cover",
                   citation = "Hurst, JM and Allen, RB. (2007)
                        The Recce method for describing New Zealand vegetation – Field protocols.
                        Landcare Research, Lincoln.",
                   codes = c("P","1","2","3", "4", "5", "6"),
                   quantifiableCodes = c("1","2","3", "4", "5", "6"),
                   breaks = c(0, 1, 5, 25, 50, 75, 100),
                   midPoints = c(0.05, 0.5, 15, 37.5, 62.5, 87.5),
                   definitions = c("Presence", "<1\%", "1-5\%","6-25\%", "26-50\%",
                                  "51-75\%", "76-100\%"))
strataDef = defineMixedStrata(name = "Recce strata",
                   description = "Standard Recce stratum definition",
                   citation = "Hurst, JM and Allen, RB. (2007)
                      The Recce method for describing New Zealand vegetation – Field protocols.
                      Landcare Research, Lincoln.",
                   heightStrataBreaks = c(0, 0.3,2.0,5, 12, 25, 50),
                   heightStrataNames = paste0("Tier ",1:6),
                   categoryStrataNames = "Tier 7",
                   categoryStrataDefinition = "Epiphytes")
x = addAggregateOrganismObservations(newVegX(), moki_tcv,
                   mapping = mapping,
                   methods = c(cover=coverscale),
                   stratumDefinition = strataDef)

# Inspect the original organism identities
head(showElementTable(x, "organismIdentity"))

y = setOriginalIdentificationConcepts(x, citationStringAll="Allen 1998")

head(showElementTable(y, "organismIdentity"))
}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
Other organism identity functions: 
\code{\link{setPreferredTaxonNomenclature}()}
}
\concept{organism identity functions}
