% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformOrdinalScale.R
\name{transformOrdinalScale}
\alias{transformOrdinalScale}
\title{Transform ordinal scale}
\usage{
transformOrdinalScale(target, method, newMethod, replaceValues = FALSE,
  verbose = TRUE)
}
\arguments{
\item{target}{The initial object of class \code{\linkS4class{VegX}} to be modified.}

\item{method}{An integer (index) or a name of an ordinal scale method.}

\item{newMethod}{An integer (index) or a name of a quantitative method existing in the initial object,
or an object of class \code{\linkS4class{VegXMethod}}.}

\item{replaceValues}{A boolean flag to indicate that values in the new scale should replace the old ones, instead of defining new measurements.
For some measurements transformations will not be possible if replacement is not forced using this flag.}

\item{verbose}{A boolean flag to indicate console output of the data transformation process.}
}
\value{
The modified object of class \code{\linkS4class{VegX}}.
}
\description{
Transforms all the values in a VegX object made using an ordinal scale into a quantitative scale
appropriate for the midpoint values of the ordinal classes.
}
\details{
The function will normally create new measurements without destroying the original ones, unless replacement is forced by setting \code{replaceValues = TRUE}.
Veg-X only allows a single measurement per observations of some kinds:
\itemize{
  \item{"diameterMeasurement" and "heightMeasurement" of indvidual organism observations.}
  \item{"heightMeasurement" of aggregate organism observations.}
  \item{"lowerLimitMeasurement" and "upperLimitMeasurement" of stratum observations.}
}
In these cases, scale transformations are not possible if \code{replaceValues = FALSE}.
}
\examples{
data(mokihinui)

# Create Veg-X document with aggregate organism observations with ordinal cover scale
taxmapping = list(plotName = "Plot", obsStartDate = "obsDate", authorTaxonName = "PreferredSpeciesName",
              stratumName = "Tier", cover = "Category")
coverscale = defineOrdinalScaleMethod(name = "Recce cover scale",
                   description = "Recce recording method by Hurst/Allen",
                   subject = "plant cover",
                   citation = "Hurst, JM and Allen, RB. (2007) The Recce method for describing New Zealand vegetation – Field protocols. Landcare Research, Lincoln.",
                   codes = c("P","1","2","3", "4", "5", "6"),
                   quantifiableCodes = c("1","2","3", "4", "5", "6"),
                   breaks = c(0, 1, 5, 25, 50, 75, 100),
                   midPoints = c(0.05, 0.5, 15, 37.5, 62.5, 87.5),
                   definitions = c("Presence", "<1\%", "1-5\%","6-25\%", "26-50\%", "51-75\%", "76-100\%"))
strataDef = defineMixedStrata(name = "Recce strata",
                              description = "Standard Recce stratum definition",
                              citation = "Hurst, JM and Allen, RB. (2007) The Recce method for describing New Zealand vegetation – Field protocols. Landcare Research, Lincoln.",
                              heightStrataBreaks = c(0, 0.3,2.0,5, 12, 25, 50),
                              heightStrataNames = paste0("Tier ",1:6),
                              categoryStrataNames = "Tier 7",
                              categoryStrataDefinition = "Epiphytes")
x = addAggregateOrganismObservations(newVegX(), tcv, "Mokihinui",
                        mapping = taxmapping,
                        methods = c(cover=coverscale),
                        stratumDefinition = strataDef)

#Add stratum observations with ordinal cover scale
mapping = list(plotName = "Plot", obsStartDate = "obsDate", stratumName = "Tier",
               cover = "CoverClass")

x = addStratumObservations(x, tier, "Mokihinui",
                        mapping = mapping,
                        methods = list(cover=coverscale),
                        stratumDefinition = strataDef)


# Transform from "Recce cover scale" to "Plant cover/\%"
percentScale = predefinedMeasurementMethod("Plant cover/\%")
y = transformOrdinalScale(x, "Recce cover scale", percentScale)

}
\seealso{
Other transform functions: \code{\link{transformQuantitativeScale}}
}
