% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegX.R
\docType{class}
\name{VegX-class}
\alias{VegX-class}
\title{Veg-X S4 class}
\description{
Implements the minimum subset of Veg-X elements for importing and
combining vegetation data
}
\section{Slots}{

\describe{
\item{\code{plots}}{list.}

\item{\code{plotObservations}}{list.}

\item{\code{taxonNames}}{list.}

\item{\code{taxonNameUsageConcepts}}{list.}

\item{\code{individualObservations}}{list.}

\item{\code{aggregatedObservations}}{list.}

\item{\code{stratumObservations}}{list.}

\item{\code{strata}}{list.}

\item{\code{individualOrganisms}}{list.}

\item{\code{attributes}}{list.}
}}

