% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegX.R
\docType{class}
\name{VegX-class}
\alias{VegX-class}
\title{S4 class for Veg-X documents}
\description{
Implements the minimum subset of Veg-X elements needed for importing and combining vegetation data in this package.
Only main elements of Veg-X are included here as lists (subelements are not specified in the class definition).
Other main elements ('taxonConcept', 'taxonDetermination', 'communityConcept', 'communityDetermination') are not yet implemented.
}
\section{Slots}{

\describe{
\item{\code{projects}}{list of research projects underpinning data collection}

\item{\code{plots}}{list of vegetation plots where measurements were made.}

\item{\code{plotObservations}}{list of vegetation plot observations.}

\item{\code{taxonNameUsageConcepts}}{list of usages of taxon names as made by the observer.}

\item{\code{strata}}{list of stratum definitions.}

\item{\code{stratumObservations}}{list of observations made on strata (e.g. overall tree cover).}

\item{\code{aggregatedObservations}}{list of observations made on aggregated organisms (e.g. plant abundance by taxa or stratum).}

\item{\code{individualOrganisms}}{list of individual organisms.}

\item{\code{individualObservations}}{list of observations made on individual organisms (e.g. diameter or height values).}

\item{\code{abioticObservations}}{list of abiotic observations made plots.}

\item{\code{vegetationObservations}}{list of observations applying to the whole stand (e.g. stand maturity)}

\item{\code{surfaceCovers}}{list of surface cover definitions.}

\item{\code{surfaceCoverObservations}}{list of surface cover observations.}

\item{\code{ancillaryObservations}}{list of ancillary observations (e.g. land use, management regime, legal protection, ...)}

\item{\code{methods}}{list of measurement methods (e.g. cover scales).}

\item{\code{attributes}}{list of attribute value description.}
}}

\examples{
showClass("VegX")

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
\code{\link{newVegX}}
}
