% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegX.R
\docType{class}
\name{VegX-class}
\alias{VegX-class}
\title{Veg-X S4 class}
\description{
Implements the minimum subset of Veg-X elements for importing and
combining vegetation data
}
\section{Slots}{

\describe{
\item{\code{projects}}{list of research projects underpinning data collection}

\item{\code{plots}}{list of vegetation plots where measurements were made.}

\item{\code{plotObservations}}{list of vegetation plot observations.}

\item{\code{taxonNames}}{list of scientific taxon names referenced in the data.}

\item{\code{taxonNameUsageConcepts}}{list of usages of taxon names.}

\item{\code{individualObservations}}{list of observations made on individual organisms (e.g. diameter or height values).}

\item{\code{aggregatedObservations}}{list of observations made on aggregated organisms (e.g. plant abundance by taxa or stratum).}

\item{\code{stratumObservations}}{list of observations made on strata (e.g. herb or shrub cover).}

\item{\code{strata}}{list of stratum definitions.}

\item{\code{individualOrganisms}}{list of individual organisms.}

\item{\code{methods}}{list of measurement methods (e.g. cover scales).}

\item{\code{attributes}}{list of attribute value description.}
}}

