% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegX.R, R/summary.VegX.R
\docType{class}
\name{VegX-class}
\alias{VegX-class}
\alias{summary,VegX-method}
\title{S4 class for Veg-X documents}
\usage{
\S4method{summary}{VegX}(object, ...)
}
\description{
Implements the minimum subset of Veg-X elements needed for importing and combining vegetation data in this package.
Only main elements of Veg-X are included here as lists (subelements are not specified in the class definition).
Other main elements ('taxonConcept', 'taxonDetermination', 'communityConcept', 'communityDetermination') are not yet implemented.
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: Generates a summary of the Veg-X object (i.e., the number of elements of each type).
}}

\section{Slots}{

\describe{
\item{\code{VegXVersion}}{A string to indicate the version of the Veg-X standard.}

\item{\code{projects}}{A list of research projects underpinning data collection.}

\item{\code{plots}}{A list of vegetation plots where measurements were made.}

\item{\code{plotObservations}}{A list of vegetation plot observations.}

\item{\code{taxonNameUsageConcepts}}{A list of usages of taxon names as made by the observer.}

\item{\code{strata}}{A list of stratum definitions.}

\item{\code{stratumObservations}}{A list of observations made on strata (e.g. overall tree cover).}

\item{\code{aggregateObservations}}{A list of observations made on aggregate organisms (e.g. plant abundance by taxa or stratum).}

\item{\code{individualOrganisms}}{A list of individual organisms.}

\item{\code{individualObservations}}{A list of observations made on individual organisms (e.g. diameter or height values).}

\item{\code{siteObservations}}{A list of observations that apply to the site (e.g., abiotic measurements, land use, management regime, legal protection, ...).}

\item{\code{communityObservations}}{list of observations applying to the whole plant community (e.g. stand age or species richness)}

\item{\code{surfaceCovers}}{A list of surface cover definitions.}

\item{\code{surfaceCoverObservations}}{list of surface cover observations.}

\item{\code{methods}}{A list of measurement methods (e.g. cover scales).}

\item{\code{attributes}}{A list of attribute value description.}
}}

\examples{
showClass("VegX")

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
\seealso{
\code{\link{newVegX}}, \code{\link{showElementTable}}
}
