% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegXMethod.R
\docType{class}
\name{VegXMethod-class}
\alias{VegXMethod-class}
\title{S4 class for Veg-X measurement methods}
\description{
S4 class for Veg-X measurement methods
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the measurement method.}

\item{\code{description}}{Description of the measurement method.}

\item{\code{citation}}{A string with the bibliographic reference for the method.}

\item{\code{attributeClass}}{Kind of attribute measured (e.g. cover).}

\item{\code{attributeType}}{Either "quantitative", "ordinal" or "qualitative".}

\item{\code{attributes}}{List of attribute values}
}}

\examples{
showClass("VegXMethod")

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
