% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegXStrata.R
\docType{class}
\name{VegXStrata-class}
\alias{VegXStrata-class}
\title{S4 class for Veg-X strata definition}
\description{
S4 class for Veg-X strata definition
}
\section{Slots}{

\describe{
\item{\code{method}}{VegXMethod.}

\item{\code{strata}}{list.}
}}

\examples{
showClass("VegXStrata")

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
