% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegXSurfaceCoverDefinition.R
\docType{class}
\name{VegXSurfaceCoverDefinition-class}
\alias{VegXSurfaceCoverDefinition-class}
\title{S4 class for Veg-X surface cover definition}
\description{
S4 class for Veg-X surface cover definition
}
\section{Slots}{

\describe{
\item{\code{method}}{VegXMethod.}

\item{\code{strata}}{list.}
}}

\examples{
showClass("VegXSurfaceCoverDefinition")

}
