% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeVegXML.R
\name{writeVegXML}
\alias{writeVegXML}
\title{Write VegX XML file}
\usage{
writeVegXML(x, file)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{VegX}}}

\item{file}{the file name to be written}
}
\value{
A string of the file written
}
\description{
Assembles and writes an XML file on the disk 
following the Veg-X XML schema standard (ver 1.6.0)
}
\examples{
\dontrun{
  target = newVegX()
  writeVegXML(target, "foo.xml")
}

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
