% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeVegXML.R
\name{writeVegXML}
\alias{writeVegXML}
\title{Write VegX XML file}
\usage{
writeVegXML(x, file, verbose = TRUE)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{VegX}}}

\item{file}{the file name to be written}

\item{verbose}{A boolean flag to indicate console output of the XML tree building process.}
}
\description{
Assembles and writes an XML file on the disk
following the Veg-X XML schema standard (ver 2.0.1)
}
\examples{
\dontrun{
  target = newVegX()
  writeVegXML(target, "foo.xml")
}

}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
