% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeVegXML.R
\name{writeVegXML}
\alias{writeVegXML}
\title{Write VegX XML file (ver 1.5.3)}
\usage{
writeVegXML(x, file)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{VegX}}}

\item{file}{the file name to be written}
}
\value{
A string of the file written
}
\description{
Write VegX XML file (ver 1.5.3)
}
\references{
Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data
}
